Built-in material allowing the user to test the package with demo data

Built-in tibbles and lists allowing the user to test the package with demo material for harmonization.

DEMO_files_harmo

Format

list

A list with 5 elements used for testing the package (data frames and lists):

DEMO_data_processing_elements - final

Data processing element without error

DEMO_data_processing_elements - with error

Data processing element containing errors

DEMO_data_processing_elements - work in progress

Data processing element in construction

DEMO_dataschema

DataSchema used, combined with data processing elements and study-specific material

pooled_harmonized_dataset

pooled harmonized dataset for testing purpose

harmonized_dossier

harmonized dossier for testing purpose

summary_var_harmo

harmonized variables summary for testing purpose

...

Examples

{

 print(DEMO_files_harmo$`dataschema - final`)

}
#> $Variables
#> # A tibble: 13 × 27
#>    index project table             name  `label:en` valueType unit  other_format
#>    <dbl> <chr>   <chr>             <chr> <chr>      <chr>     <chr> <lgl>       
#>  1     1 DEMO    harmonized_study… adm_… Unique id… text      NA    NA          
#>  2     2 DEMO    harmonized_study… adm_… Indicator… text      NA    NA          
#>  3     3 DEMO    harmonized_study… adm_… Indicator… text      NA    NA          
#>  4     4 DEMO    harmonized_study… sdc_… Participa… integer   years NA          
#>  5     5 DEMO    harmonized_study… sdc_… Gender of… integer   NA    NA          
#>  6     6 DEMO    harmonized_study… phy_… participa… decimal   cm    NA          
#>  7     7 DEMO    harmonized_study… phy_… participa… decimal   kg    NA          
#>  8     8 DEMO    harmonized_study… phy_… participa… decimal   kg/m  NA          
#>  9     9 DEMO    harmonized_study… rep_… whether t… integer   NA    NA          
#> 10    10 DEMO    harmonized_study… rep_… whether t… integer   NA    NA          
#> 11    11 DEMO    harmonized_study… lsb_… whether t… integer   NA    NA          
#> 12    12 DEMO    harmonized_study… lsb_… whether t… integer   NA    NA          
#> 13    13 DEMO    harmonized_study… lsb_… participa… integer   NA    NA          
#> # ℹ 19 more variables: individual_entity <lgl>, time_period <lgl>,
#> #   source <lgl>, informant <lgl>, time_collection <lgl>, wording <lgl>,
#> #   measures <lgl>, procedures <lgl>, instructions <lgl>, comments <lgl>,
#> #   `Mlstr_additional::Source` <lgl>, `Mlstr_additional::Target` <lgl>,
#> #   `Mlstr_area::1` <chr>, `Mlstr_area::1.term` <chr>,
#> #   `Mlstr_area::1.scale` <chr>, `Mlstr_area::2` <lgl>,
#> #   `Mlstr_area::2.term` <lgl>, `Mlstr_area::3` <lgl>, …
#> 
#> $Categories
#> # A tibble: 13 × 6
#>    project table                            variable    `label:en`  name missing
#>    <chr>   <chr>                            <chr>       <chr>      <dbl>   <dbl>
#>  1 DEMO    harmonized_study_DEMO_table_DEMO sdc_gender  Male           1       0
#>  2 DEMO    harmonized_study_DEMO_table_DEMO sdc_gender  Female         2       0
#>  3 DEMO    harmonized_study_DEMO_table_DEMO rep_preg_e… never pre…     0       0
#>  4 DEMO    harmonized_study_DEMO_table_DEMO rep_preg_e… pregnant …     1       0
#>  5 DEMO    harmonized_study_DEMO_table_DEMO rep_preg_c… currently…     0       0
#>  6 DEMO    harmonized_study_DEMO_table_DEMO rep_preg_c… not curre…     1       0
#>  7 DEMO    harmonized_study_DEMO_table_DEMO lsb_smo_ev… never smo…     0       0
#>  8 DEMO    harmonized_study_DEMO_table_DEMO lsb_smo_ev… smoked on…     1       0
#>  9 DEMO    harmonized_study_DEMO_table_DEMO lsb_smo_cu… currently…     0       0
#> 10 DEMO    harmonized_study_DEMO_table_DEMO lsb_smo_cu… not curre…     1       0
#> 11 DEMO    harmonized_study_DEMO_table_DEMO lsb_smo_st… never smo…     0       0
#> 12 DEMO    harmonized_study_DEMO_table_DEMO lsb_smo_st… former sm…     1       0
#> 13 DEMO    harmonized_study_DEMO_table_DEMO lsb_smo_st… current s…     2       0
#>