Validate and coerce any object as data processing element

Confirms that the input object is a valid data processing element, and return it as a data processing element with the appropriate mlstr_class attribute. This function mainly helps validate input within other functions of the package but could be used to check if a data processing element is valid.

as_data_procelem(object)

Arguments

object

A potential data processing element(list of tibble) to be coerced.

Value

A list of tibble(s), each of them identifying a data processing elements.

Details

A data processing element must be a list of at least one data frame or data frame extension (e.g. a tibble) named 'dataProcess'. These elements are the centralised decision making to perform harmonization of each study-specific dataset. The tibble must contain certain columns which participate to the process, including the 'rule_category' and 'harmo_rule'. In addition, the data processing element may follow Maelstrom research standards, and its content can be evaluated accordingly, such as naming convention restriction, columns like 'valueType', and 'harmo status details', and/or any harmonization taxonomy provided.

Examples

{

# You can use our demonstration files to run examples
library(madshapR)

as_data_procelem(DEMO_files$`data_processing_elements - final`)
}
#> $dataProcessing
#> # A tibble: 52 × 19
#>    index ss_da…¹ datas…² ss_ac…³ ss_ta…⁴ ss_va…⁵ ss_fo…⁶ ss_ti…⁷ value…⁸ ss_ti…⁹
#>    <dbl> <chr>   <chr>   <chr>   <chr>   <chr>   <lgl>   <lgl>   <chr>   <lgl>  
#>  1     1 harmon… adm_un… MELBOU… datase… id      NA      NA      text    NA     
#>  2     1 harmon… adm_un… MELBOU… datase… id      NA      NA      text    NA     
#>  3     1 harmon… adm_un… PARIS   datase… ID      NA      NA      text    NA     
#>  4     1 harmon… adm_un… TOKYO   datase… part_id NA      NA      text    NA     
#>  5     2 harmon… adm_st… MELBOU… datase… __BLAN… NA      NA      text    NA     
#>  6     2 harmon… adm_st… MELBOU… datase… __BLAN… NA      NA      text    NA     
#>  7     2 harmon… adm_st… PARIS   datase… __BLAN… NA      NA      text    NA     
#>  8     2 harmon… adm_st… TOKYO   datase… __BLAN… NA      NA      text    NA     
#>  9     3 harmon… adm_ye… MELBOU… datase… __BLAN… NA      NA      text    NA     
#> 10     3 harmon… adm_ye… MELBOU… datase… __BLAN… NA      NA      text    NA     
#> # … with 42 more rows, 9 more variables: ss_source_information <lgl>,
#> #   ss_informant <lgl>, ss_keywords <lgl>, comment <lgl>, harmo_status <lgl>,
#> #   harmo_status_detail <lgl>, rule_category <chr>, harmo_rule <chr>,
#> #   comment_internal <lgl>, and abbreviated variable names
#> #   ¹​ss_dataschema_table, ²​dataschema_variable, ³​ss_acronym, ⁴​ss_table,
#> #   ⁵​ss_variables, ⁶​ss_format, ⁷​ss_time_period, ⁸​valueType, ⁹​ss_time_collection
#> 
#> attr(,"Mlstr::class")
#> [1] "data_procelem"