Validate and coerce any object as Dataschema

Confirms that the input object is a valid Dataschema, and return it as a Dataschema with the appropriate mlstr_class attribute. This function mainly helps validate input within other functions of the package but could be used to check if a Dataschema is valid.

as_dataschema(object, as_mlstr_dataschema = FALSE)

Arguments

object

A potential Dataschema (list of tibble) to be coerced.

as_mlstr_dataschema

Whether the output Dataschema has a simple Dataschema structure or not (meaning has a Maelstrom Dataschema structure, compatible with Maelstrom ecosystem such as Opal environment). FALSE by default.

Value

A list of tibble(s), each of them identifying a Dataschema.

Details

The Dataschema must be a list of at least one or two data frame or data frame extension (e.g. a tibble) named 'Variables' and 'Categories' (if any), representing meta data of an associated harmonized study. The Dataschema is the core list of variables to be generate using study-specific variables. In addition, the Dataschema may follow Maelstrom research standards, and its content can be evaluated accordingly, such as naming convention restriction, columns like valueType' and 'label(:xx)', and/or any harmonization taxonomy provided.

Examples

{

# You can use our demonstration files to run examples
library(madshapR)

as_dataschema(DEMO_files$`dataschema - final`)
}
#> $Variables
#> # A tibble: 13 × 11
#>    name         typeof index project table label…¹ value…² unit  Mlstr…³ Mlstr…⁴
#>    <chr>        <chr>  <chr> <chr>   <chr> <chr>   <chr>   <chr> <chr>   <chr>  
#>  1 adm_unique_… chara… 1     DEMO    harm… Unique… text    NA    ADM     Identi…
#>  2 adm_study    chara… 2     DEMO    harm… Indica… text    NA    ADM     Questi…
#>  3 adm_year_dce chara… 3     DEMO    harm… Indica… text    NA    ADM     Questi…
#>  4 sdc_age      integ… 4     DEMO    harm… Partic… integer years SDC     Age    
#>  5 sdc_gender   integ… 5     DEMO    harm… Gender… integer NA    SDC     Sex    
#>  6 phy_height   double 6     DEMO    harm… partic… decimal cm    PME     Anthro…
#>  7 phy_weight   double 7     DEMO    harm… partic… decimal kg    PME     Anthro…
#>  8 phy_bmi      double 8     DEMO    harm… partic… decimal kg/m  PME     Anthro…
#>  9 rep_preg_ev… integ… 9     DEMO    harm… whethe… integer NA    REP     Pregna…
#> 10 rep_preg_cu… integ… 10    DEMO    harm… whethe… integer NA    REP     Pregna…
#> 11 lsb_smo_ever integ… 11    DEMO    harm… whethe… integer NA    LSB     Tobacco
#> 12 lsb_smo_curr integ… 12    DEMO    harm… whethe… integer NA    LSB     Tobacco
#> 13 lsb_smo_sta… integ… 13    DEMO    harm… partic… integer NA    LSB     Tobacco
#> # … with 1 more variable: `Mlstr_area::1.scale` <chr>, and abbreviated variable
#> #   names ¹​`label:en`, ²​valueType, ³​`Mlstr_area::1`, ⁴​`Mlstr_area::1.term`
#> 
#> $Categories
#> # A tibble: 13 × 7
#>    variable       name  labels project table                     label…¹ missing
#>    <chr>          <chr> <chr>  <chr>   <chr>                     <chr>   <chr>  
#>  1 sdc_gender     1     1      DEMO    harmonized_study_DEMO_ta… Male    0      
#>  2 sdc_gender     2     2      DEMO    harmonized_study_DEMO_ta… Female  0      
#>  3 rep_preg_ever  0     0      DEMO    harmonized_study_DEMO_ta… never … 0      
#>  4 rep_preg_ever  1     1      DEMO    harmonized_study_DEMO_ta… pregna… 0      
#>  5 rep_preg_curr  0     0      DEMO    harmonized_study_DEMO_ta… curren… 0      
#>  6 rep_preg_curr  1     1      DEMO    harmonized_study_DEMO_ta… not cu… 0      
#>  7 lsb_smo_ever   0     0      DEMO    harmonized_study_DEMO_ta… never … 0      
#>  8 lsb_smo_ever   1     1      DEMO    harmonized_study_DEMO_ta… smoked… 0      
#>  9 lsb_smo_curr   0     0      DEMO    harmonized_study_DEMO_ta… curren… 0      
#> 10 lsb_smo_curr   1     1      DEMO    harmonized_study_DEMO_ta… not cu… 0      
#> 11 lsb_smo_status 0     0      DEMO    harmonized_study_DEMO_ta… never … 0      
#> 12 lsb_smo_status 1     1      DEMO    harmonized_study_DEMO_ta… former… 0      
#> 13 lsb_smo_status 2     2      DEMO    harmonized_study_DEMO_ta… curren… 0      
#> # … with abbreviated variable name ¹​`label:en`
#> 
#> attr(,"Mlstr::class")
#> [1] "dataschema"