Validate and coerce any object as a DataSchema with Maelstrom restrictions

Validates the input object as a valid DataSchema compliant with formats used in Maelstrom Research ecosystem, including Opal, and returns it with the appropriate harmonizR::class attribute. This function mainly helps validate input within other functions of the package but could be used to check if an object is valid for use in a function.

as_dataschema_mlstr(object)

Arguments

object

A potential Maelstrom formatted Dataschema to be coerced.

Value

A list of tibble(s), each of them identifying the DataSchema.

Details

A DataSchema defines the harmonized variables to be generated, representing meta data of an associated harmonized dossier. It must be a list of data frame like objects with elements named 'Variables' (required) and 'Categories' (if any). The 'Variables' element must contain at least the 'name' column, and the 'Categories' element must contain at least the 'variable' and 'name' columns to be usable in any function. To be considered as a minimum workable DataSchema, in 'Variables' the 'name' column must also have unique and non-null entries, and in 'Categories' the combination of 'variable' and 'name' columns must also be unique'.

Examples

{

# You can use our demonstration files to run examples
library(madshapR) # import DEMO_files

as_dataschema_mlstr(DEMO_files$`dataschema - final`)

}
#> $Variables
#> # A tibble: 13 × 10
#>    name           `label:en` valueType index project table unit  `Mlstr_area::1`
#>    <chr>          <chr>      <chr>     <chr> <chr>   <chr> <chr> <chr>          
#>  1 adm_unique_id  Unique id… text      1     DEMO    harm… NA    ADM            
#>  2 adm_study      Indicator… text      2     DEMO    harm… NA    ADM            
#>  3 adm_year_dce   Indicator… text      3     DEMO    harm… NA    ADM            
#>  4 sdc_age        Participa… integer   4     DEMO    harm… years SDC            
#>  5 sdc_gender     Gender of… integer   5     DEMO    harm… NA    SDC            
#>  6 phy_height     participa… decimal   6     DEMO    harm… cm    PME            
#>  7 phy_weight     participa… decimal   7     DEMO    harm… kg    PME            
#>  8 phy_bmi        participa… decimal   8     DEMO    harm… kg/m  PME            
#>  9 rep_preg_ever  whether t… integer   9     DEMO    harm… NA    REP            
#> 10 rep_preg_curr  whether t… integer   10    DEMO    harm… NA    REP            
#> 11 lsb_smo_ever   whether t… integer   11    DEMO    harm… NA    LSB            
#> 12 lsb_smo_curr   whether t… integer   12    DEMO    harm… NA    LSB            
#> 13 lsb_smo_status participa… integer   13    DEMO    harm… NA    LSB            
#> # ℹ 2 more variables: `Mlstr_area::1.term` <chr>, `Mlstr_area::1.scale` <chr>
#> 
#> $Categories
#> # A tibble: 13 × 6
#>    variable       name  `label:en`                         missing project table
#>    <chr>          <chr> <chr>                              <lgl>   <chr>   <chr>
#>  1 sdc_gender     1     Male                               FALSE   DEMO    harm…
#>  2 sdc_gender     2     Female                             FALSE   DEMO    harm…
#>  3 rep_preg_ever  0     never pregnant                     FALSE   DEMO    harm…
#>  4 rep_preg_ever  1     pregnant once or more              FALSE   DEMO    harm…
#>  5 rep_preg_curr  0     currently pregnant                 FALSE   DEMO    harm…
#>  6 rep_preg_curr  1     not currently pregnant             FALSE   DEMO    harm…
#>  7 lsb_smo_ever   0     never smoked                       FALSE   DEMO    harm…
#>  8 lsb_smo_ever   1     smoked one pack of cigarette or m… FALSE   DEMO    harm…
#>  9 lsb_smo_curr   0     currently smoker                   FALSE   DEMO    harm…
#> 10 lsb_smo_curr   1     not currently smoker               FALSE   DEMO    harm…
#> 11 lsb_smo_status 0     never smoker                       FALSE   DEMO    harm…
#> 12 lsb_smo_status 1     former smoker                      FALSE   DEMO    harm…
#> 13 lsb_smo_status 2     current smoker                     FALSE   DEMO    harm…
#> 
#> attr(,"madshapR::class")
#> [1] "data_dict_mlstr"
#> attr(,"harmonizR::class")
#> [1] "Dataschema_mlstr"