Validate and coerce any object as an harmonized dossier

Validates the input object as a valid harmonized dossier and coerces it with the appropriate harmonizR::class attribute. This function mainly helps validate input within other functions of the package but could be used to check if an object is valid for use in a function.

as_harmonized_dossier(
  object,
  dataschema = NULL,
  data_proc_elem = NULL,
  harmonized_col_id = NULL
)

Arguments

object

A potential harmonized dossier to be coerced.

dataschema

A list of tibble(s) representing meta data of an associated harmonized dossier.

data_proc_elem

A tibble, identifying the input data processing elements.

harmonized_col_id

A character string identifying the name of the column present in every dataset as identifier of the dataset.

Value

A list of tibble(s), each of them identifying the harmonized dataset.

Details

A harmonized dossier must be a named list containing at least one data frame or data frame extension (e.g. a tibble), each of them being harmonized dataset(s). It is generally the product of applying harmonization processing to a dossier object. The name of each tibble will be use as the reference name of the dataset. A harmonized dossier has four attributes : harmonizR::class which is ""harmonized_dossier"" ; harmonizR::Dataschema (provided by user) ; harmonizR::data processing elements ; harmonizR::harmonized_col_id (provided by user) which refers to the column in each dataset which identifies unique combination observation/dataset. This id column name is the same across the dataset(s), the DataSchema and the data processing elements (created by using 'id_creation') and is used to initiate the process of harmonization.

A DataSchema defines the harmonized variables to be generated, representing meta data of an associated harmonized dossier. It must be a list of data frame like objects with elements named 'Variables' (required) and 'Categories' (if any). The 'Variables' element must contain at least the 'name' column, and the 'Categories' element must contain at least the 'variable' and 'name' columns to be usable in any function. To be considered as a minimum workable DataSchema, in 'Variables' the 'name' column must also have unique and non-null entries, and in 'Categories' the combination of 'variable' and 'name' columns must also be unique'.

A data processing element contains the rules and metadata that will be used to perform harmonization of input datasets in accordance with the DataSchema. It must be a data-frame or data-frame extension (e.g. a tibble) and it must contain certain columns which participate to the process, including the dataschema_variable, ss-table,ss_variables, Mlstr::rule_category and Mlstr::algorithm. The mandatory first processing element must be "id_creation" in Mlstr::rule_category followed by the name of the column taken as identifier of each dataset to initiate the process of harmonization.

Examples

{

# You can use our demonstration files to run examples
library(madshapR) # import DEMO_files
library(fabR)     # silently_run
library(lubridate)

dataset_MELBOURNE_1 <- DEMO_files$dataset_MELBOURNE_1
dataset_MELBOURNE_2 <- DEMO_files$dataset_MELBOURNE_2
dossier <- dossier_create(list(dataset_MELBOURNE_1, dataset_MELBOURNE_2))

dataschema <- DEMO_files$`dataschema - final`

data_proc_elem <- 
  DEMO_files$`data_processing_elements - final`$`dataProcessing`
  
# perform harmonization 
harmonized_dossier <- 
  silently_run(harmo_process(dossier,dataschema,data_proc_elem))

as_harmonized_dossier(
  harmonized_dossier,dataschema,data_proc_elem,'adm_unique_id')
  
}
#> Warning: package 'fabR' was built under R version 4.2.3
#> 
#> Attaching package: 'lubridate'
#> The following objects are masked from 'package:base':
#> 
#>     date, intersect, setdiff, union
#> $dataset_MELBOURNE_1
#> # A tibble: 19 × 13
#>    adm_unique_id adm_study adm_year_dce sdc_age sdc_gender phy_height phy_weight
#>    <chr>         <chr>     <chr>          <int> <int+lbl>       <dbl>      <dbl>
#>  1 377943        MELBOURNE 2007              52 2 [Female]         NA         NA
#>  2 497013        MELBOURNE 2007              49 1 [Male]           NA         NA
#>  3 927676        MELBOURNE 2007              43 1 [Male]           NA         NA
#>  4 995667        MELBOURNE 2007              59 2 [Female]         NA         NA
#>  5 21829         MELBOURNE 2007              40 2 [Female]         NA         NA
#>  6 209432        MELBOURNE 2007              47 1 [Male]           NA         NA
#>  7 272983        MELBOURNE 2007              NA 2 [Female]         NA         NA
#>  8 580632        MELBOURNE 2007              53 2 [Female]         NA         NA
#>  9 304624        MELBOURNE 2007              35 2 [Female]         NA         NA
#> 10 637551        MELBOURNE 2007              40 1 [Male]           NA         NA
#> 11 279817        MELBOURNE 2007              41 1 [Male]           NA         NA
#> 12 235415        MELBOURNE 2007              34 2 [Female]         NA         NA
#> 13 373673        MELBOURNE 2007              48 2 [Female]         NA         NA
#> 14 485098        MELBOURNE 2007              43 2 [Female]         NA         NA
#> 15 299427        MELBOURNE 2007              NA 1 [Male]           NA         NA
#> 16 854073        MELBOURNE 2007              41 1 [Male]           NA         NA
#> 17 197666        MELBOURNE 2007              33 2 [Female]         NA         NA
#> 18 130327        MELBOURNE 2007              57 2 [Female]         NA         NA
#> 19 220050        MELBOURNE 2007              50 2 [Female]         NA         NA
#> # ℹ 6 more variables: phy_bmi <dbl>, rep_preg_ever <int+lbl>,
#> #   rep_preg_curr <int+lbl>, lsb_smo_ever <int+lbl>, lsb_smo_curr <int+lbl>,
#> #   lsb_smo_status <int+lbl>
#> 
#> $dataset_MELBOURNE_2
#> # A tibble: 19 × 13
#>    adm_unique_id adm_study adm_year_dce sdc_age sdc_gender phy_height phy_weight
#>    <chr>         <chr>     <chr>          <int> <int+lbl>       <dbl>      <dbl>
#>  1 377943        MELBOURNE 2007              NA NA                 NA         NA
#>  2 497013        MELBOURNE 2007              NA NA                 NA         NA
#>  3 927676        MELBOURNE 2007              NA NA                 NA         NA
#>  4 995667        MELBOURNE 2007              NA NA                 NA         NA
#>  5 21829         MELBOURNE 2007              NA NA                 NA         NA
#>  6 209432        MELBOURNE 2007              NA NA                 NA         NA
#>  7 272983        MELBOURNE 2007              NA NA                 NA         NA
#>  8 580632        MELBOURNE 2007              NA NA                 NA         NA
#>  9 304624        MELBOURNE 2007              NA NA                 NA         NA
#> 10 637551        MELBOURNE 2007              NA NA                 NA         NA
#> 11 279817        MELBOURNE 2007              NA NA                 NA         NA
#> 12 235415        MELBOURNE 2007              NA NA                 NA         NA
#> 13 373673        MELBOURNE 2007              NA NA                 NA         NA
#> 14 485098        MELBOURNE 2007              NA NA                 NA         NA
#> 15 299427        MELBOURNE 2007              NA NA                 NA         NA
#> 16 854073        MELBOURNE 2007              NA NA                 NA         NA
#> 17 197666        MELBOURNE 2007              NA NA                 NA         NA
#> 18 130327        MELBOURNE 2007              NA NA                 NA         NA
#> 19 220050        MELBOURNE 2007              NA NA                 NA         NA
#> # ℹ 6 more variables: phy_bmi <dbl>, rep_preg_ever <int+lbl>,
#> #   rep_preg_curr <int+lbl>, lsb_smo_ever <int+lbl>, lsb_smo_curr <int+lbl>,
#> #   lsb_smo_status <int+lbl>
#> 
#> attr(,"harmonizR::class")
#> [1] "harmonized_dossier"
#> attr(,"harmonizR::Dataschema")
#> attr(,"harmonizR::Dataschema")$Variables
#> # A tibble: 13 × 10
#>    name           `label:en` valueType index project table unit  `Mlstr_area::1`
#>    <chr>          <chr>      <chr>     <chr> <chr>   <chr> <chr> <chr>          
#>  1 adm_unique_id  Unique id… text      1     DEMO    harm… NA    ADM            
#>  2 adm_study      Indicator… text      2     DEMO    harm… NA    ADM            
#>  3 adm_year_dce   Indicator… text      3     DEMO    harm… NA    ADM            
#>  4 sdc_age        Participa… integer   4     DEMO    harm… years SDC            
#>  5 sdc_gender     Gender of… integer   5     DEMO    harm… NA    SDC            
#>  6 phy_height     participa… decimal   6     DEMO    harm… cm    PME            
#>  7 phy_weight     participa… decimal   7     DEMO    harm… kg    PME            
#>  8 phy_bmi        participa… decimal   8     DEMO    harm… kg/m  PME            
#>  9 rep_preg_ever  whether t… integer   9     DEMO    harm… NA    REP            
#> 10 rep_preg_curr  whether t… integer   10    DEMO    harm… NA    REP            
#> 11 lsb_smo_ever   whether t… integer   11    DEMO    harm… NA    LSB            
#> 12 lsb_smo_curr   whether t… integer   12    DEMO    harm… NA    LSB            
#> 13 lsb_smo_status participa… integer   13    DEMO    harm… NA    LSB            
#> # ℹ 2 more variables: `Mlstr_area::1.term` <chr>, `Mlstr_area::1.scale` <chr>
#> 
#> attr(,"harmonizR::Dataschema")$Categories
#> # A tibble: 13 × 6
#>    variable       name  `label:en`                         missing project table
#>    <chr>          <chr> <chr>                              <lgl>   <chr>   <chr>
#>  1 lsb_smo_curr   0     currently smoker                   FALSE   DEMO    harm…
#>  2 lsb_smo_curr   1     not currently smoker               FALSE   DEMO    harm…
#>  3 lsb_smo_ever   0     never smoked                       FALSE   DEMO    harm…
#>  4 lsb_smo_ever   1     smoked one pack of cigarette or m… FALSE   DEMO    harm…
#>  5 lsb_smo_status 0     never smoker                       FALSE   DEMO    harm…
#>  6 lsb_smo_status 1     former smoker                      FALSE   DEMO    harm…
#>  7 lsb_smo_status 2     current smoker                     FALSE   DEMO    harm…
#>  8 rep_preg_curr  0     currently pregnant                 FALSE   DEMO    harm…
#>  9 rep_preg_curr  1     not currently pregnant             FALSE   DEMO    harm…
#> 10 rep_preg_ever  0     never pregnant                     FALSE   DEMO    harm…
#> 11 rep_preg_ever  1     pregnant once or more              FALSE   DEMO    harm…
#> 12 sdc_gender     1     Male                               FALSE   DEMO    harm…
#> 13 sdc_gender     2     Female                             FALSE   DEMO    harm…
#> 
#> attr(,"harmonizR::Dataschema")attr(,"madshapR::class")
#> [1] "data_dict_mlstr"
#> attr(,"harmonizR::Dataschema")attr(,"harmonizR::class")
#> [1] "Dataschema_mlstr"
#> attr(,"harmonizR::data processing elements")
#> # A tibble: 52 × 19
#>    index dataschema_variable valueType ss_table            ss_variables
#>  * <dbl> <chr>               <chr>     <chr>               <chr>       
#>  1     1 adm_unique_id       text      dataset_MELBOURNE_1 id          
#>  2     1 adm_unique_id       text      dataset_MELBOURNE_2 id          
#>  3     1 adm_unique_id       text      dataset_PARIS       ID          
#>  4     1 adm_unique_id       text      dataset_TOKYO       part_id     
#>  5     2 adm_study           text      dataset_MELBOURNE_1 __BLANK__   
#>  6     2 adm_study           text      dataset_MELBOURNE_2 __BLANK__   
#>  7     2 adm_study           text      dataset_PARIS       __BLANK__   
#>  8     2 adm_study           text      dataset_TOKYO       __BLANK__   
#>  9     3 adm_year_dce        text      dataset_MELBOURNE_1 __BLANK__   
#> 10     3 adm_year_dce        text      dataset_MELBOURNE_2 __BLANK__   
#> # ℹ 42 more rows
#> # ℹ 14 more variables: `Mlstr::rule_category` <chr>, `Mlstr::algorithm` <chr>,
#> #   `Mlstr::comment` <lgl>, `Mlstr::status` <lgl>,
#> #   `Mlstr::status_detail` <lgl>, ss_dataschema_table <chr>, ss_acronym <chr>,
#> #   ss_format <lgl>, ss_time_period <lgl>, ss_time_collection <lgl>,
#> #   ss_source_information <lgl>, ss_informant <lgl>, ss_keywords <lgl>,
#> #   comment_internal <lgl>
#> attr(,"harmonizR::harmonized_col_id")
#> [1] "adm_unique_id"