Generate a quality assessment report of the DataSchema

Assesses the content and structure of a DataSchema and reports potential issues to facilitate the assessment of input data. The report can be used to help assess data structure, presence of fields, coherence across elements, and taxonomy or data dictionary formats. This report is compatible with Excel and can be exported as an Excel spreadsheet.

dataschema_evaluate(dataschema, taxonomy = NULL)

Arguments

dataschema

A list of tibble(s) representing meta data of an * associated harmonized dossier.

taxonomy

A tibble identifying the scheme used for variables classification.

Value

A list of tibbles of report for the DataSchema.

Details

A DataSchema defines the harmonized variables to be generated, representing meta data of an associated harmonized dossier. It must be a list of data frame like objects with elements named 'Variables' (required) and 'Categories' (if any). The 'Variables' element must contain at least the 'name' column, and the 'Categories' element must contain at least the 'variable' and 'name' columns to be usable in any function. To be considered as a minimum workable DataSchema, in 'Variables' the 'name' column must also have unique and non-null entries, and in 'Categories' the combination of 'variable' and 'name' columns must also be unique.

A taxonomy is classification scheme that can be defined for variable attributes. If defined, a taxonomy must be a data frame like object. It must be compatible with (and is generally extracted from) an Opal environment. To work with certain functions, a valid taxonomy must contain at least the columns 'taxonomy', 'vocabulary', and 'terms'. In addition, the taxonomy may follow Maelstrom research taxonomy, and its content can be evaluated accordingly, such as naming convention restriction, tagging elements, or scales, which are specific to Maelstrom Research. In this particular case, the tibble must also contain 'vocabulary_short', 'taxonomy_scale', 'vocabulary_scale' and 'term_scale' to work with some specific functions.

Examples

{

library(madshapR)
library(janitor)
dataschema <- 
  DEMO_files_harmo$`dataschema - final` %>%
  data_dict_filter("name == 'adm_unique_id'") %>%
  lapply(function(x) remove_empty(x,which = 'cols'))
  
dataschema_evaluate(dataschema)

}
#> 
#> Attaching package: 'janitor'
#> The following objects are masked from 'package:stats':
#> 
#>     chisq.test, fisher.test
#> - DATA DICTIONARY ASSESSMENT: dataschema --------------
#>     Assess the standard adequacy of naming
#>     Assess the uniqueness of variable names
#>     Assess the presence of possible duplicated columns
#>     Assess the presence of empty rows in the data dictionary
#>     Assess the presence of empty columns in the data dictionary
#>     Assess the completion of `label(:xx)` column in 'Variables'
#>     Assess the `valueType` column in 'Variables'
#>     Generate report
#> 
#>     The data dictionary contains no error/warning.
#> 
#>   - WARNING MESSAGES (if any): --------------------------------------------
#> 
#> $`Study-specific Dataschema summary`
#> # A tibble: 1 × 8
#>   index name          `label:en`         valueType project table `Mlstr_area::1`
#>   <int> <chr>         <chr>              <chr>     <chr>   <chr> <chr>          
#> 1     1 adm_unique_id Unique identifica… text      DEMO    harm… ADM            
#> # ℹ 1 more variable: `Mlstr_area::1.term` <chr>
#>