Search an index of files

Searches in mlstr_file_index R object (tibble) based on pattern and other query options and provides a table where all the files in a specified folder and corresponding to the query are listed (recursively). If no index tibble is provided, the function creates one from the working directory.

file_index_search(
  index = tibble(),
  file_path = "",
  file_name = "",
  extension = "",
  file_type = "",
  .fs_tree = TRUE
)

Arguments

index
The index (tibble) of a folder with file locations and metadata, either previously generated by file_index_create() or created from folder.

file_path
A character string specifying a file path to search by. Can be the full string or substring (matching by regex is supported)

file_name
A character string a file name to search by. Can be the full string or substring (matching by regex is supported).

extension
A character string a file extention to search by. Can be the full string or substring (matching by regex is supported).

file_type
A character string a file type to search by. Can be the full string or substring (matching by regex is supported).

.fs_tree
If TRUE, return the file tree of the query.

Value

A tibble with indexed information for files matching the query.

Examples

if (FALSE) {
# use case 1: create index of files in one folder
index_DEMO <- file_index_create(folder = "TEST_demo")

# use case 3: subset index matching a specific file_path query
file_index_search(index = index_DEMO, file_path = "data_processing_elements")

# use case 4: subset index matching a specific file_type query
file_index_search(index = index_DEMO, file_name = "study_MELBOURNE_1")

}