Generate a summary report for a harmonized dataset

This function generates a summary report in an bookdown (html-based) document showing descriptive statistics for each variable in a harmonized dataset to facilitate assessment of harmonized data. Statistics and figures are generated according to four types of variables: numerical, text, categorical, and dual (numerical variables with categories, e.g., “Do not know”, “Do not want to answer”). The variable type is automatically detected from the dataset and its data dictionary. The output can be editable (using plotly library) or static (using ggplot2 library).

generate_harmonized_study_bookdown_report(
  dataschema,
  data_proc_elem,
  harmonized_dataset,
  group_by = "adm_study",
  to = paste0("bookdown_report/", as.character(substitute(harmonized_dataset))),
  out = "plotly",
  .keep_files = FALSE
)

Arguments

dataschema
A tibble identifying the DataSchema to use for harmonization

data_proc_elem
A character string or tibble R object identifying the (annotated) data processing elements file

harmonized_dataset
A character string or tibble identifying the input dataset.

group_by
A character string of one column in the dataset that can be taken as a grouping column. the vizual element will be grouped and displayed by this column. Default is “adm_study” (following Maelstrom standard naming).

to
A character string identifying the folder path where the bookdown report will be saved.

out
parameter that specifies the graphical outputs expected in the report: can be either 'ggplot2' or 'plotly'. gglot2 renders static plots, plotly dynamic plots.

.keep_files
Boolean value to say whether to keep the R-markdown files.

Value

A bookdown folder containing files in the specified output folder (current folder/docs by default). To open the file in browser, open "index.html"

Details

Must provide a harmonized dataset and DataSchema in Maelstrom formats.

Examples

if (FALSE) {
# use case 1: report of harmonized_study_DEMO_table_DEMO.
generate_harmonized_study_bookdown_report(
  harmonized_dataset = harmonized_study_DEMO_table_DEMO,
  dataschema = Harmo_dataschema,
  data_proc_elem = annotated_data_proc_elem,
  to = "demo_files/reports",
  # group_by = "adm_study"
)

}