Generate a summary report bookdown for a study dataset

This function generates a summary report in an bookdown (html-based) document showing descriptive statistics for each variable in an individual study dataset to facilitate assessment of input data. Statistics and figures are generated according to four types of variables: numerical, text, categorical, and dual (open field variables with not open categories, e.g., “Do not know”, “Do not want to answer”). The variable type is automatically detected from the dataset and its data dictionary. The output can be editable (using plotly library) or static (using ggplot2 library).

generate_individual_study_bookdown_report(
  dataset,
  data_dict,
  group_by = NULL,
  to = paste0("bookdown_report/", as.character(substitute(dataset))),
  out = "plotly",
  .keep_files = FALSE
)

Arguments

dataset
A character string or tibble R object identifying the input dataset (in Maelstrom format). Maelstrom-format dataset has the entity (usually participant) ID as the first column.

data_dict
A tibble identifying the data dictionary (in Maelstrom format) associated with the dataset. Automatically generated if not provided.

group_by
A character string of one column in the dataset that can be taken as a grouping column. the vizual element will be grouped and displayed by this column.

to
A character string identifying the folder path where the bookdown report will be saved.

out
parameter that specifies the graphical outputs expected in the report: can be either 'ggplot2' or 'plotly'. gglot2 renders static plots, plotly dynamic plots.

.keep_files
Boolean value to say whether to keep the R-markdown files.

Value

A bookdown folder containing files in the specified output folder (current folder/docs by default). To open the file in browser, open "index.html"

Details

Must provide dataset and data dictionary in Maelstrom formats. If the data dictionary is not provided, a basic one is being created during the process.

Examples

if (FALSE) {

# Create index of files in one folder and read the files
index_demo_files <- mlstr_file_index_create(folder = "demo_files")
mlstr_file_index_read(index = index_demo_files, file_name = "study_TOKYO")
mlstr_file_index_read(index = index_demo_files, file_name = "dd_TOKYO")

# use case 1: report of demo dataset TOKYO
generate_individual_study_bookdown_report(
  dataset = study_TOKYO,
  data_dict = dd_TOKYO_format_maelstrom_tagged,
  to = "demo_files/reports/TOKYO")

# use case 2: report of demo dataset TOKYO, grouped by gndr
generate_individual_study_bookdown_report(
  dataset = study_TOKYO,
  data_dict = dd_TOKYO_format_maelstrom_tagged,
  to = "demo_files/reports/TOKYO_gndr",group_by = "gndr",out = "ggplot2")

# re-index your files to include new files created
index_demo_files <- mlstr_file_index_create(folder = "demo_files")

# read the book down
mlstr_file_index_read(index_demo_files,file_path = "demo_files/reports/TOKYO_gndr/docs/index.html")
}