Generate harmonized dataset(s) and annotated data processing elements file

Reads the Dataschema and data processing elements R objects to generate harmonized dataset(s) and annotated data processing elements file with harmonization status and errors. The function uses the Dataschema and data processing elements specifications to process input study-specific variables into study harmonized variables. Documentation of each data processing action is generated in the console to support the identification of errors and correction of the data processing elements file and objects, as needed. An annotated data processing elements report is also produced providing harmonization statuses (complete/impossible) for each Dataschema variable in each study-specific dataset, which can be used to create a summary of the harmonization potential of the Dataschema variables across study-specific datasets.

harmo_process(study, dataschema = NULL, data_procelem)

Arguments

study

List of tibble, each of them being study specific datasets.

dataschema

A list of tibble(s) representing meta data of an associated harmonized study.

data_procelem

A list of tibble(s), identifying the input data processing elements.

Value

A list of tibbles of each harmonized dataset that has been harmonized from study specific datataset.

Details

The Dataschema must be a list of at least one or two data frame or data frame extension (e.g. a tibble) named 'Variables' and 'Categories' (if any), representing meta data of an associated harmonized study. The Dataschema is the core list of variables to be generate using study-specific variables. In addition, the Dataschema may follow Maelstrom research standards, and its content can be evaluated accordingly, such as naming convention restriction, columns like valueType' and 'label(:xx)', and/or any harmonization taxonomy provided.

An harmonized study must be a named list containing at least one data frame or data frame extension (e.g. a tibble), each of them being harmonized datasets. The name of each tibble will be use as the reference name of the dataset. If the harmonization process contains errors, the tibbles and csv files will not be generated to avoid any version confusion. The user can correct their data processing element file and re-run the process. If the user identified errors and want to correct them later, they can specify undetermined' in the column 'harmo_rule' of their problematic data processing element(s). Such data processing element will not be processed.

A data processing element must be a list of at least one data frame or data frame extension (e.g. a tibble) named 'dataProcess'. These elements are the centralised decision making to perform harmonization of each study-specific dataset. The tibble must contain certain columns which participate to the process, including the 'rule_category' and 'harmo_rule'. In addition, the data processing element may follow Maelstrom research standards, and its content can be evaluated accordingly, such as naming convention restriction, columns like 'valueType', and 'harmo status details', and/or any harmonization taxonomy provided.

Examples

{

library(madshapR)
library(lubridate)
library(dplyr)

dataset_MELBOURNE_1 <- DEMO_files$dataset_MELBOURNE_1
dataset_MELBOURNE_2 <- DEMO_files$dataset_MELBOURNE_2
study <-
  study_create(list(dataset_MELBOURNE_1, dataset_MELBOURNE_2)) %>%
  lapply(. %>% slice(1:2))

dataschema <-
  list(Variables =
         DEMO_files$`dataschema - final`$Variables[
           c('name','label:en','valueType')])

###### Example 1 : the Data processing elements contains errors. The
# function returns a tibble with the errors annotated does not produce
# the harmonized datasets.
harmo_process(study,dataschema,
  data_procelem = DEMO_files$`data_processing_elements - with error`)

###### Example 2 : the Data processing elements contains errors. The
# function returns a tibble with the errors annotated does not produce
# the harmonized datasets.#'
harmo_process(study,dataschema,
  data_procelem = DEMO_files$`data_processing_elements - final`)

}
#> 
#> Attaching package: ‘lubridate’
#> The following objects are masked from ‘package:base’:
#> 
#>     date, intersect, setdiff, union
#> 
#> Attaching package: ‘dplyr’
#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag
#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union
#> - DATA PROCESSING ELEMENTS: ------------------------------------------------
#> 
#> --harmonization of : dataset_MELBOURNE_1 ------------------------------
#>     processing 1/13 : adm_unique_id     id created
#>     processing 2/13 : adm_study         complete
#>     processing 3/13 : adm_year_dce      complete
#>     processing 4/13 : sdc_age           complete
#>     processing 5/13 : sdc_gender        **ERROR**
#>     processing 6/13 : phy_height        impossible
#>     processing 7/13 : phy_weight        impossible
#>     processing 8/13 : phy_bmi           complete
#>     processing 9/13 : rep_preg_ever     impossible
#>     processing 10/13 : rep_preg_curr    impossible
#>     processing 11/13 : lsb_smo_ever     impossible
#>     processing 12/13 : lsb_smo_curr     impossible
#>     processing 13/13 : lsb_smo_status   impossible
#> 
#> --harmonization of : dataset_MELBOURNE_2 ------------------------------
#>     processing 1/13 : adm_unique_id     id created
#>     processing 2/13 : adm_study         complete
#>     processing 3/13 : adm_year_dce      complete
#>     processing 4/13 : sdc_age           impossible
#>     processing 5/13 : sdc_gender        impossible
#>     processing 6/13 : phy_height        impossible
#>     processing 7/13 : phy_weight        impossible
#>     processing 8/13 : phy_bmi           impossible
#>     processing 9/13 : rep_preg_ever     impossible
#>     processing 10/13 : rep_preg_curr    complete
#>     processing 11/13 : lsb_smo_ever     complete
#>     processing 12/13 : lsb_smo_curr     **ERROR**
#>     processing 13/13 : lsb_smo_status   complete
#> 
#> 
#> ------------------------------------------------------------------------------
#> 
#> Your harmonization process contains errors. The tables affected have not been
#> generated to avoid any version confusion. Correct your data processing element
#> file and re-run the process.
#> 
#> 
#> Useful tip: If you identified errors and want to correct them later, you can specify
#> 'undetermined' in the column 'harmo_rule' of your problematic data processing
#> element(s). Such data processing element will not be processed.
#> - DATA PROCESSING ELEMENTS: ------------------------------------------------
#> 
#> --harmonization of : dataset_MELBOURNE_1 ------------------------------
#>     processing 1/13 : adm_unique_id     id created
#>     processing 2/13 : adm_study         complete
#>     processing 3/13 : adm_year_dce      complete
#>     processing 4/13 : sdc_age           complete
#>     processing 5/13 : sdc_gender        complete
#>     processing 6/13 : phy_height        impossible
#>     processing 7/13 : phy_weight        impossible
#>     processing 8/13 : phy_bmi           complete
#>     processing 9/13 : rep_preg_ever     impossible
#>     processing 10/13 : rep_preg_curr    impossible
#>     processing 11/13 : lsb_smo_ever     impossible
#>     processing 12/13 : lsb_smo_curr     impossible
#>     processing 13/13 : lsb_smo_status   impossible
#> 
#> --harmonization of : dataset_MELBOURNE_2 ------------------------------
#>     processing 1/13 : adm_unique_id     id created
#>     processing 2/13 : adm_study         complete
#>     processing 3/13 : adm_year_dce      complete
#>     processing 4/13 : sdc_age           impossible
#>     processing 5/13 : sdc_gender        impossible
#>     processing 6/13 : phy_height        impossible
#>     processing 7/13 : phy_weight        impossible
#>     processing 8/13 : phy_bmi           impossible
#>     processing 9/13 : rep_preg_ever     impossible
#>     processing 10/13 : rep_preg_curr    complete
#>     processing 11/13 : lsb_smo_ever     complete
#>     processing 12/13 : lsb_smo_curr     complete
#>     processing 13/13 : lsb_smo_status   complete
#> 
#> 
#> - CREATION OF STUDY SPECIFIC DATASCHEMA: ----------------------------------
#> 
#> dataset_MELBOURNE_1 : done
#> dataset_MELBOURNE_2 : done
#> 
#> 
#> ------------------------------------------------------------------------------
#> 
#> Your harmonization is done. Please check if everything worked correctly.
#> 
#> - WARNING MESSAGES (if any): ----------------------------------------------
#> 
#> $dataset_MELBOURNE_1
#> # A tibble: 2 × 13
#>   adm_unique_id adm_st…¹ adm_y…² sdc_age sdc_g…³ phy_h…⁴ phy_w…⁵ phy_bmi rep_p…⁶
#>   <chr>         <chr>    <chr>     <int>   <int>   <dbl>   <dbl>   <dbl>   <int>
#> 1 377943        MELBOUR… 2007         52       2      NA      NA    22.1      NA
#> 2 497013        MELBOUR… 2007         49       1      NA      NA    18.5      NA
#> # … with 4 more variables: rep_preg_curr <int>, lsb_smo_ever <int>,
#> #   lsb_smo_curr <int>, lsb_smo_status <int>, and abbreviated variable names
#> #   ¹​adm_study, ²​adm_year_dce, ³​sdc_gender, ⁴​phy_height, ⁵​phy_weight,
#> #   ⁶​rep_preg_ever
#> 
#> $dataset_MELBOURNE_2
#> # A tibble: 2 × 13
#>   adm_unique_id adm_st…¹ adm_y…² sdc_age sdc_g…³ phy_h…⁴ phy_w…⁵ phy_bmi rep_p…⁶
#>   <chr>         <chr>    <chr>     <int>   <int>   <dbl>   <dbl>   <dbl>   <int>
#> 1 377943        MELBOUR… 2007         NA      NA      NA      NA      NA      NA
#> 2 497013        MELBOUR… 2007         NA      NA      NA      NA      NA      NA
#> # … with 4 more variables: rep_preg_curr <int>, lsb_smo_ever <int>,
#> #   lsb_smo_curr <int>, lsb_smo_status <int>, and abbreviated variable names
#> #   ¹​adm_study, ²​adm_year_dce, ³​sdc_gender, ⁴​phy_height, ⁵​phy_weight,
#> #   ⁶​rep_preg_ever
#> 
#> attr(,"Mlstr::class")
#> [1] "study"