Generate a web-based bookdown visual report of a harmonized dossier

Generates a visual report for a dataset in an HTML bookdown document. The report provides figures and descriptive statistics for each variable to facilitate the assessment of input data. Statistics and figures are generated according to variable data type. The report can be used to help assess data structure, coherence across elements, and taxonomy or data dictionary formats. The summaries and figures provide additional information about variable distributions and descriptive statistics. The charts and tables are produced based on their data type. The variable can be grouped using group_by parameter, which is a (categorical) column in the dataset. The user may need to use as.factor() in this context. To fasten the process (and allow recycling object in a workflow) the user can feed the function with a .summary_var, which is the output of the function dataset_summarize() of the column(s) col and group_by. The summary must have the same parameters to operate.

harmonized_dossier_visualize(
  harmonized_dossier,
  to,
  taxonomy = NULL,
  .summary_var = NULL,
  .keep_files = TRUE
)

Arguments

harmonized_dossier

List of tibble(s), each of them being harmonized dataset.

to

A character string identifying the folder path where the bookdown report will be saved.

taxonomy

A tibble identifying the scheme used for variables classification.

.summary_var

A list which is the summary of the variables.

.keep_files

whether to keep the R-markdown files. TRUE by default. (used for internal processes and programming).

Value

A bookdown folder containing files in the specified output folder. To open the file in browser, open 'docs/index.html'. Or use open_visual_report()

Details

A harmonized dossier must be a named list containing at least one data frame or data frame extension (e.g. a tibble), each of them being harmonized dataset(s). It is generally the product of applying harmonization processing to a dossier object. The name of each tibble will be use as the reference name of the dataset. A harmonized dossier has four attributes : harmonizR::class which is ""harmonized_dossier"" ; harmonizR::Dataschema (provided by user) ; harmonizR::data processing elements ; harmonizR::harmonized_col_id (provided by user) which refers to the column in each dataset which identifies unique combination observation/dataset. This id column name is the same across the dataset(s), the DataSchema and the data processing elements (created by using 'id_creation') and is used to initiate the process of harmonization.

A taxonomy is classification scheme that can be defined for variable attributes. If defined, a taxonomy must be a data frame like object. It must be compatible with (and is generally extracted from) an Opal environment. To work with certain functions, a valid taxonomy must contain at least the columns 'taxonomy', 'vocabulary', and 'terms'. In addition, the taxonomy may follow Maelstrom research taxonomy, and its content can be evaluated accordingly, such as naming convention restriction, tagging elements, or scales, which are specific to Maelstrom Research. In this particular case, the tibble must also contain 'vocabulary_short', 'taxonomy_scale', 'vocabulary_scale' and 'term_scale' to work with some specific functions.

See also

Examples

{

# You can use our demonstration files to run examples
library(madshapR) # import DEMO_files
library(fabR) # silently_run
library(lubridate)

dataset_MELBOURNE_1 <- DEMO_files$dataset_MELBOURNE_1
dataset_MELBOURNE_2 <- DEMO_files$dataset_MELBOURNE_2
dossier <- dossier_create(list(dataset_MELBOURNE_1, dataset_MELBOURNE_2))

dataschema <- DEMO_files$`dataschema - final`

data_proc_elem <- 
  DEMO_files$`data_processing_elements - final`$`dataProcessing`
  
# perform harmonization 
harmonized_dossier <- 
  silently_run(harmo_process(dossier,dataschema,data_proc_elem))

# generate viz report
# Silencing the process, remove silently_run() to see the full process in 
# the console. 
to = tempdir()
silently_run(harmonized_dossier_visualize(harmonized_dossier, to=tempdir()))
 
# To open the file in browser, open 'to/docs/index.html'. 
# Or use [open_visual_report()]
 
open_visual_report(to)

}
#> 
  |                                                        
  |                                                  |   0%
  |                                                        
  |                                                  |   1%                    
  |                                                        
  |.                                                 |   2% (unnamed-chunk-1)                      
  |                                                        
  |..                                                |   3% (unnamed-chunk-2)  
  |                                                        
  |..                                                |   4%                    
  |                                                        
  |..                                                |   5% (unnamed-chunk-3)  
  |                                                        
  |...                                               |   6%                     (unnamed-chunk-4)  
  |                                                        
  |....                                              |   7%                    
  |                                                        
  |....                                              |   8% (unnamed-chunk-5)  
  |                                                        
  |....                                              |   9%                    
  |                                                        
  |.....                                             |  10% (unnamed-chunk-6)                      
  |                                                        
  |......                                            |  11% (unnamed-chunk-7)  
  |                                                        
  |......                                            |  12%                    
  |                                                        
  |......                                            |  13% (figures-plot12-2) 
  |                                                        
  |.......                                           |  14%                     (unnamed-chunk-8)  
  |                                                        
  |........                                          |  15%                    
  |                                                        
  |........                                          |  16% (unnamed-chunk-9)  
  |                                                        
  |........                                          |  17%                    
  |                                                        
  |.........                                         |  18% (unnamed-chunk-10)                     
  |                                                        
  |..........                                        |  19% (figures-plot12-3) 
  |                                                        
  |..........                                        |  20%                    
  |                                                        
  |..........                                        |  21% (unnamed-chunk-11) 
  |                                                        
  |...........                                       |  22%                     (unnamed-chunk-12) 
  |                                                        
  |............                                      |  23%                    
  |                                                        
  |............                                      |  24% (unnamed-chunk-13) 
  |                                                        
  |............                                      |  25%                    
  |                                                        
  |.............                                     |  26% (figures-plot12-4)                     
  |                                                        
  |..............                                    |  27% (unnamed-chunk-14) 
  |                                                        
  |..............                                    |  28%                    
  |                                                        
  |..............                                    |  29% (unnamed-chunk-15) 
  |                                                        
  |...............                                   |  30%                     (unnamed-chunk-16) 
  |                                                        
  |................                                  |  31%                    
  |                                                        
  |................                                  |  32% (figures-plot12-5) 
  |                                                        
  |................                                  |  33%                    
  |                                                        
  |.................                                 |  34% (unnamed-chunk-17)                     
  |                                                        
  |..................                                |  35% (unnamed-chunk-18) 
  |                                                        
  |..................                                |  36%                    
  |                                                        
  |..................                                |  37% (unnamed-chunk-19) 
  |                                                        
  |...................                               |  38%                     (unnamed-chunk-20) 
  |                                                        
  |....................                              |  39%                    
  |                                                        
  |....................                              |  40% (figures-plot12-6) 
  |                                                        
  |....................                              |  41%                    
  |                                                        
  |.....................                             |  42% (unnamed-chunk-21)                     
  |                                                        
  |......................                            |  43% (unnamed-chunk-22) 
  |                                                        
  |......................                            |  44%                    
  |                                                        
  |......................                            |  45% (unnamed-chunk-23) 
  |                                                        
  |.......................                           |  46%                     (figures-plot12-7) 
  |                                                        
  |........................                          |  47%                    
  |                                                        
  |........................                          |  48% (unnamed-chunk-24) 
  |                                                        
  |........................                          |  49%                    
  |                                                        
  |.........................                         |  50% (unnamed-chunk-25)                     
  |                                                        
  |..........................                        |  51% (unnamed-chunk-26) 
  |                                                        
  |..........................                        |  52%                    
  |                                                        
  |..........................                        |  53% (figures-plot12-8) 
  |                                                        
  |...........................                       |  54%                     (unnamed-chunk-27) 
  |                                                        
  |............................                      |  55%                    
  |                                                        
  |............................                      |  56% (unnamed-chunk-28) 
  |                                                        
  |............................                      |  57%                    
  |                                                        
  |.............................                     |  58% (unnamed-chunk-29)                     
  |                                                        
  |..............................                    |  59% (figures-plot12-9) 
  |                                                        
  |..............................                    |  60%                    
  |                                                        
  |..............................                    |  61% (unnamed-chunk-30) 
  |                                                        
  |...............................                   |  62%                     (unnamed-chunk-31) 
  |                                                        
  |................................                  |  63%                    
  |                                                        
  |................................                  |  64% (unnamed-chunk-32) 
  |                                                        
  |................................                  |  65%                    
  |                                                        
  |.................................                 |  66% (unnamed-chunk-33)                     
  |                                                        
  |..................................                |  67% (figures-plot12-10)
  |                                                        
  |..................................                |  68%                    
  |                                                        
  |..................................                |  69% (unnamed-chunk-34) 
  |                                                        
  |...................................               |  70%                     (unnamed-chunk-35) 
  |                                                        
  |....................................              |  71%                    
  |                                                        
  |....................................              |  72% (unnamed-chunk-36) 
  |                                                        
  |....................................              |  73%                    
  |                                                        
  |.....................................             |  74% (unnamed-chunk-37)                     
  |                                                        
  |......................................            |  75% (figures-plot12-11)
  |                                                        
  |......................................            |  76%                    
  |                                                        
  |......................................            |  77% (unnamed-chunk-38) 
  |                                                        
  |.......................................           |  78%                     (unnamed-chunk-39) 
  |                                                        
  |........................................          |  79%                    
  |                                                        
  |........................................          |  80% (unnamed-chunk-40) 
  |                                                        
  |........................................          |  81%                    
  |                                                        
  |.........................................         |  82% (unnamed-chunk-41)                     
  |                                                        
  |..........................................        |  83% (figures-plot12-12)
  |                                                        
  |..........................................        |  84%                    
  |                                                        
  |..........................................        |  85% (unnamed-chunk-42) 
  |                                                        
  |...........................................       |  86%                     (unnamed-chunk-43) 
  |                                                        
  |............................................      |  87%                    
  |                                                        
  |............................................      |  88% (unnamed-chunk-44) 
  |                                                        
  |............................................      |  89%                    
  |                                                        
  |.............................................     |  90% (unnamed-chunk-45)                     
  |                                                        
  |..............................................    |  91% (figures-plot12-13)
  |                                                        
  |..............................................    |  92%                    
  |                                                        
  |..............................................    |  93% (unnamed-chunk-46) 
  |                                                        
  |...............................................   |  94%                     (unnamed-chunk-47) 
  |                                                        
  |................................................  |  95%                    
  |                                                        
  |................................................  |  96% (unnamed-chunk-48) 
  |                                                        
  |................................................  |  97%                    
  |                                                        
  |................................................. |  98% (unnamed-chunk-49)                     
  |                                                        
  |..................................................|  99% (figures-plot12-14)
  |                                                        
  |..................................................| 100%                    
                                                                                                              
#> "C:/Program Files/RStudio/resources/app/bin/quarto/bin/tools/pandoc" +RTS -K512m -RTS bookdownproj.knit.md --to html4 --from markdown+autolink_bare_uris+tex_math_single_backslash --output bookdownproj.html --lua-filter "C:\Users\guill\AppData\Local\R\win-library\4.2\bookdown\rmarkdown\lua\custom-environment.lua" --lua-filter "C:\Users\guill\AppData\Local\R\win-library\4.2\rmarkdown\rmarkdown\lua\pagebreak.lua" --lua-filter "C:\Users\guill\AppData\Local\R\win-library\4.2\rmarkdown\rmarkdown\lua\latex-div.lua" --lua-filter "C:\Users\guill\AppData\Local\R\win-library\4.2\rmarkdown\rmarkdown\lua\anchor-sections.lua" --metadata-file "C:\Users\guill\AppData\Local\Temp\RtmpO8XQ3G\file39549f52a8d" --wrap preserve --standalone --section-divs --table-of-contents --toc-depth 3 --template "C:\Users\guill\AppData\Local\R\win-library\4.2\bookdown\templates\gitbook.html" --highlight-style pygments --number-sections --css style.css --mathjax --include-in-header "C:\Users\guill\AppData\Local\Temp\RtmpO8XQ3G\rmarkdown-str39543d6358c1.html"