Generate an Excel spreadsheet report of a harmonization project

Generates an Excel spreadsheet report for a harmonized study, showing descriptive statistics for each harmonized variable to facilitate the assessment of input data. Statistics are generated according to their valueType. This report can be used to assist the user in the assessment of the data structure, fields investigation (mandatory or not), coherence across elements and taxonomy, or standard evaluation. The summary associated provides dataset composition, with observation distribution and descriptive statistics.

harmonized_study_summarise(harmonized_study, valueType_guess = TRUE)

Arguments

harmonized_study

List of tibble(s), each of them being harmonized datasets. an associated harmonized dataset. Automatically generated if not provided.

valueType_guess

Whether the output should include a more accurate valueType that could be applied to the dataset. TRUE by default.

Value

A list of tibbles of report for each harmonized dataset.

Details

The Dataschema must be a list of at least one or two data frame or data frame extension (e.g. a tibble) named 'Variables' and 'Categories' (if any), representing meta data of an associated harmonized study. The Dataschema is the core list of variables to be generate using study-specific variables. In addition, the Dataschema may follow Maelstrom research standards, and its content can be evaluated accordingly, such as naming convention restriction, columns like 'valueType' and 'label(:xx)', and/or any harmonization taxonomy provided.

An harmonized study must be a named list containing at least one data frame or data frame extension (e.g. a tibble), each of them being harmonized datasets. The name of each tibble will be use as the reference name of the dataset. If the harmonization process contains errors, the tibbles and csv files will not be generated to avoid any version confusion. The user can correct their data processing element file and re-run the process. If the user identified errors and want to correct them later, they can specify 'undetermined' in the column 'harmo_rule' of their problematic data processing element(s). Such data processing element will not be processed.

A data processing element must be a list of at least one data frame or data frame extension (e.g. a tibble) named 'dataProcess'. These elements are the centralised decision making to perform harmonization of each study-specific dataset. The tibble must contain certain columns which participate to the process, including the 'rule_category' and 'harmo_rule'. In addition, the data processing element may follow Maelstrom research standards, and its content can be evaluated accordingly, such as naming convention restriction, columns like 'valueType', and 'harmo status details', and/or any harmonization taxonomy provided.

Examples

{

# You can use our demonstration files to run examples
library(madshapR)
library(lubridate)
library(dplyr)

dataset_MELBOURNE_1 <- DEMO_files$dataset_MELBOURNE_1
dataset_MELBOURNE_2 <- DEMO_files$dataset_MELBOURNE_2
study <-
  study_create(list(dataset_MELBOURNE_1, dataset_MELBOURNE_2)) %>%
  lapply(. %>% slice(1:2))

dataschema <-
  list(Variables =
         DEMO_files$`dataschema - final`$Variables[
           c('name','label:en','valueType')])

data_procelem <- DEMO_files$`data_processing_elements - final`
harmonized_study <- harmo_process(study,dataschema, data_procelem)

harmonized_study_summarise(harmonized_study)

}
#> - DATA PROCESSING ELEMENTS: ------------------------------------------------
#> 
#> --harmonization of : dataset_MELBOURNE_1 ------------------------------
#>     processing 1/13 : adm_unique_id     id created
#>     processing 2/13 : adm_study         complete
#>     processing 3/13 : adm_year_dce      complete
#>     processing 4/13 : sdc_age           complete
#>     processing 5/13 : sdc_gender        complete
#>     processing 6/13 : phy_height        impossible
#>     processing 7/13 : phy_weight        impossible
#>     processing 8/13 : phy_bmi           complete
#>     processing 9/13 : rep_preg_ever     impossible
#>     processing 10/13 : rep_preg_curr    impossible
#>     processing 11/13 : lsb_smo_ever     impossible
#>     processing 12/13 : lsb_smo_curr     impossible
#>     processing 13/13 : lsb_smo_status   impossible
#> 
#> --harmonization of : dataset_MELBOURNE_2 ------------------------------
#>     processing 1/13 : adm_unique_id     id created
#>     processing 2/13 : adm_study         complete
#>     processing 3/13 : adm_year_dce      complete
#>     processing 4/13 : sdc_age           impossible
#>     processing 5/13 : sdc_gender        impossible
#>     processing 6/13 : phy_height        impossible
#>     processing 7/13 : phy_weight        impossible
#>     processing 8/13 : phy_bmi           impossible
#>     processing 9/13 : rep_preg_ever     impossible
#>     processing 10/13 : rep_preg_curr    complete
#>     processing 11/13 : lsb_smo_ever     complete
#>     processing 12/13 : lsb_smo_curr     complete
#>     processing 13/13 : lsb_smo_status   complete
#> 
#> 
#> - CREATION OF STUDY SPECIFIC DATASCHEMA: ----------------------------------
#> 
#> dataset_MELBOURNE_1 : done
#> dataset_MELBOURNE_2 : done
#> 
#> 
#> ------------------------------------------------------------------------------
#> 
#> Your harmonization is done. Please check if everything worked correctly.
#> 
#> - WARNING MESSAGES (if any): ----------------------------------------------
#> 
#> - STUDY SUMMARY: -------------------------------------------------------
#> - DATA DICTIONARY ASSESSMENT: data_dict --------------
#>     Assess the standard adequacy of naming
#>     Assess the uniqueness of variable names
#>     Assess the presence of possible duplicated columns
#>     Assess the presence of empty rows in the data dictionary
#>     Assess the presence of columns in the data dictionary
#>     Assess the completion of `label(:xx)` column in 'Variables'
#>     Assess the `valueType` column in 'Variables'
#>     Generate report
#> 
#>   - WARNING MESSAGES (if any): --------------------------------------------
#> 
#> - DATASET ASSESSMENT: dataset --------------------------
#>     Assess the standard adequacy of naming
#>     Assess the presence of variable names both in dataset and data dictionary
#>     Assess the presence of possible duplicated variable in the dataset
#>     Assess the presence of duplicated participants in the dataset
#>     Assess the presence of unique value columns in dataset
#>     Assess the presence of empty rows in the data dictionary
#>     Assess the presence all NA(s) of columns in the data dictionary
#>     Assess the presence of categories not in the data dictionary
#>     Assess the `valueType` comparison in dataset and data dictionary
#>     Generate report
#> 
#>   - WARNING MESSAGES (if any): -------------------------------------------------
#>     
#> - DATASET SUMMARIZE: dataset_MELBOURNE_1 --------------------------
#>     Summarize the data type of each variable across the dataset
#>     Summarise information for all variables
#>     Summarise information for text variables
#>     Summarise information for date variables
#> Warning: Problem while computing date type variables due to ambiguous format.
#> They will be analysed as text variables
#> Useful tip:Use dataset_evaluate(data) to get an assessment of your data.
#>     Summarise information for numerical variables
#>     Summarise information for categorical variables
#>     Summarise global information (Overview)
#>     Generate report
#> - DATA DICTIONARY ASSESSMENT: data_dict --------------
#>     Assess the standard adequacy of naming
#>     Assess the uniqueness of variable names
#>     Assess the presence of possible duplicated columns
#>     Assess the presence of empty rows in the data dictionary
#>     Assess the presence of columns in the data dictionary
#>     Assess the completion of `label(:xx)` column in 'Variables'
#>     Assess the `valueType` column in 'Variables'
#>     Generate report
#> 
#>   - WARNING MESSAGES (if any): --------------------------------------------
#> 
#> - DATASET ASSESSMENT: dataset --------------------------
#>     Assess the standard adequacy of naming
#>     Assess the presence of variable names both in dataset and data dictionary
#>     Assess the presence of possible duplicated variable in the dataset
#>     Assess the presence of duplicated participants in the dataset
#>     Assess the presence of unique value columns in dataset
#>     Assess the presence of empty rows in the data dictionary
#>     Assess the presence all NA(s) of columns in the data dictionary
#>     Assess the presence of categories not in the data dictionary
#>     Assess the `valueType` comparison in dataset and data dictionary
#>     Generate report
#> 
#>   - WARNING MESSAGES (if any): -------------------------------------------------
#>     
#> - DATASET SUMMARIZE: dataset_MELBOURNE_2 --------------------------
#>     Summarize the data type of each variable across the dataset
#>     Summarise information for all variables
#>     Summarise information for text variables
#>     Summarise information for date variables
#> Warning: Problem while computing date type variables due to ambiguous format.
#> They will be analysed as text variables
#> Useful tip:Use dataset_evaluate(data) to get an assessment of your data.
#>     Summarise information for numerical variables
#>     Summarise information for categorical variables
#>     Summarise global information (Overview)
#>     Generate report
#> - DATA DICTIONARY ASSESSMENT: data_dict --------------
#>     Assess the standard adequacy of naming
#>     Assess the uniqueness of variable names
#>     Assess the presence of possible duplicated columns
#>     Assess the presence of empty rows in the data dictionary
#>     Assess the presence of columns in the data dictionary
#>     Assess the completion of `label(:xx)` column in 'Variables'
#>     Assess the `valueType` column in 'Variables'
#>     Generate report
#> 
#>   - WARNING MESSAGES (if any): --------------------------------------------
#> 
#> - DATASET ASSESSMENT: dataset --------------------------
#>     Assess the standard adequacy of naming
#>     Assess the presence of variable names both in dataset and data dictionary
#>     Assess the presence of possible duplicated variable in the dataset
#>     Assess the presence of duplicated participants in the dataset
#>     Assess the presence of unique value columns in dataset
#>     Assess the presence of empty rows in the data dictionary
#>     Assess the presence all NA(s) of columns in the data dictionary
#>     Assess the presence of categories not in the data dictionary
#>     Assess the `valueType` comparison in dataset and data dictionary
#>     Generate report
#> 
#>   - WARNING MESSAGES (if any): -------------------------------------------------
#>     
#> - DATASET SUMMARIZE: Mlstr::pooled_harmonized_dataset --------------------------
#>     Summarize the data type of each variable across the dataset
#>     Summarise information for all variables
#>     Summarise information for text variables
#>     Summarise information for date variables
#> Warning: Problem while computing date type variables due to ambiguous format.
#> They will be analysed as text variables
#> Useful tip:Use dataset_evaluate(data) to get an assessment of your data.
#>     Summarise information for numerical variables
#>     Summarise information for categorical variables
#>     Summarise global information (Overview)
#>     Generate report
#> $dataset_MELBOURNE_1
#> $dataset_MELBOURNE_1$`Overview of harmonization`
#> # A tibble: 13 × 2
#>    `Quality control of study specific dataset` ` `                  
#>    <chr>                                       <chr>                
#>  1 "Date"                                      "2023-03-17"         
#>  2 "Name of the dataset"                       "dataset_MELBOURNE_1"
#>  3 "    Participant identifier"                "adm_unique_id"      
#>  4 "    Variables"                             " "                  
#>  5 "        Total number of variables"         "13"                 
#>  6 "    Data type in dictionnary (valueType)"  " "                  
#>  7 "        Nb. text variables"                "1"                  
#>  8 "        Nb. date variables"                "0"                  
#>  9 "        Nb. numerical variables"           "12"                 
#> 10 "        Nb. categorical variables"         "0"                  
#> 11 "    Rows"                                  " "                  
#> 12 "        Total number of observations"      "2"                  
#> 13 "        Nb. unique participants"           "2"                  
#> 
#> $dataset_MELBOURNE_1$`Study specific Dataschema summary`
#> # A tibble: 13 × 7
#>    index name           `label:en`               value…¹ Mlstr…² Mlstr…³ Mlstr…⁴
#>    <chr> <chr>          <chr>                    <chr>   <chr>   <chr>   <chr>  
#>  1 1     adm_unique_id  Unique identification c… text    comple… "id"    id_cre…
#>  2 2     adm_study      Indicator of the survey… text    comple… "\"MEL… paste  
#>  3 3     adm_year_dce   Indicator of the survey… text    comple… "2007"  paste  
#>  4 4     sdc_age        Participant's age at ti… integer comple… "recod… recode 
#>  5 5     sdc_gender     Gender of the participa… integer comple… "direc… direct…
#>  6 6     phy_height     participant's height.    decimal imposs… "impos… imposs…
#>  7 7     phy_weight     participant's weight.    decimal imposs… "impos… imposs…
#>  8 8     phy_bmi        participant's BMI calcu… decimal comple… "direc… direct…
#>  9 9     rep_preg_ever  whether the participant… integer imposs… "impos… imposs…
#> 10 10    rep_preg_curr  whether the participant… integer imposs… "impos… imposs…
#> 11 11    lsb_smo_ever   whether the participant… integer imposs… "impos… imposs…
#> 12 12    lsb_smo_curr   whether the participant… integer imposs… "impos… imposs…
#> 13 13    lsb_smo_status participant status rega… integer imposs… "impos… imposs…
#> # … with abbreviated variable names ¹​valueType, ²​`Mlstr::harmo_status`,
#> #   ³​`Mlstr::algorithm`, ⁴​`Mlstr::rule_category`
#> 
#> $dataset_MELBOURNE_1$`Study specific Dataschema assessement`
#> # A tibble: 1 × 4
#>   sheet     name_col name_var `Quality assessment comment`                      
#>   <chr>     <chr>    <chr>    <chr>                                             
#> 1 Variables Mlstr    NA       [INFO] - Possible duplicated columns: Mlstr::algo…
#> 
#> $dataset_MELBOURNE_1$`Harmonized Dataset assessement`
#> # A tibble: 11 × 4
#>    name_var       `Quality assessment comment`        value suggestion
#>    <chr>          <chr>                               <chr> <chr>     
#>  1 adm_study      [INFO] - Unique value in the column NA    NA        
#>  2 adm_unique_id  [INFO] - refined valueType proposed text  integer   
#>  3 adm_year_dce   [INFO] - Unique value in the column NA    NA        
#>  4 adm_year_dce   [INFO] - refined valueType proposed text  integer   
#>  5 lsb_smo_curr   [INFO] - Empty column               NA    NA        
#>  6 lsb_smo_ever   [INFO] - Empty column               NA    NA        
#>  7 lsb_smo_status [INFO] - Empty column               NA    NA        
#>  8 phy_height     [INFO] - Empty column               NA    NA        
#>  9 phy_weight     [INFO] - Empty column               NA    NA        
#> 10 rep_preg_curr  [INFO] - Empty column               NA    NA        
#> 11 rep_preg_ever  [INFO] - Empty column               NA    NA        
#> 
#> $dataset_MELBOURNE_1$`Variables summary (all)`
#> # A tibble: 13 × 15
#>    index in da…¹ name_…² label…³ Data …⁴ Estim…⁵ Actua…⁶ Categ…⁷ Quali…⁸ categ…⁹
#>    <chr>         <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>  
#>  1 1             adm_un… Unique… text    integer text    NA      [INFO]… no     
#>  2 2             adm_st… Indica… text    text    text    NA      [INFO]… no     
#>  3 3             adm_ye… Indica… text    integer text    NA      [INFO]… no     
#>  4 4             sdc_age Partic… integer integer integer NA      [INFO]… no     
#>  5 5             sdc_ge… Gender… integer integer integer NA      [INFO]… no     
#>  6 6             phy_he… partic… decimal decimal decimal NA      [INFO]… no     
#>  7 7             phy_we… partic… decimal decimal decimal NA      [INFO]… no     
#>  8 8             phy_bmi partic… decimal decimal decimal NA      [INFO]… no     
#>  9 9             rep_pr… whethe… integer integer integer NA      [INFO]… no     
#> 10 10            rep_pr… whethe… integer integer integer NA      [INFO]… no     
#> 11 11            lsb_sm… whethe… integer integer integer NA      [INFO]… no     
#> 12 12            lsb_sm… whethe… integer integer integer NA      [INFO]… no     
#> 13 13            lsb_sm… partic… integer integer integer NA      [INFO]… no     
#> # … with 6 more variables: `Total number of observations` <dbl>,
#> #   `Nb. distinct values` <int>, `% total Valid values` <dbl>, `% NA` <dbl>,
#> #   `% Valid categorical values (if applicable)` <dbl>,
#> #   `% Missing categorical values (if applicable)` <dbl>, and abbreviated
#> #   variable names ¹​`index in data dict.`, ²​name_var, ³​`label:en`,
#> #   ⁴​`Data Dictionary valueType`, ⁵​`Estimated dataset valueType`,
#> #   ⁶​`Actual dataset valueType`, ⁷​`Categories in data dictionary`, …
#> 
#> $dataset_MELBOURNE_1$`Text variable summary`
#> # A tibble: 1 × 17
#>   index in dat…¹ name_…² label…³ Data …⁴ Estim…⁵ Actua…⁶ Categ…⁷ Quali…⁸ categ…⁹
#>   <chr>          <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>  
#> 1 2              adm_st… Indica… text    text    text    NA      [INFO]… no     
#> # … with 8 more variables: `Total number of observations` <dbl>,
#> #   `Nb. distinct values` <int>, `% total Valid values` <dbl>, `% NA` <dbl>,
#> #   `% Valid categorical values (if applicable)` <dbl>,
#> #   `% Missing categorical values (if applicable)` <dbl>,
#> #   `Most seen value(s)` <chr>, `Rarest value(s)` <chr>, and abbreviated
#> #   variable names ¹​`index in data dict.`, ²​name_var, ³​`label:en`,
#> #   ⁴​`Data Dictionary valueType`, ⁵​`Estimated dataset valueType`, …
#> 
#> $dataset_MELBOURNE_1$`Numerical variable summary`
#> # A tibble: 12 × 21
#>    index in da…¹ name_…² label…³ Data …⁴ Estim…⁵ Actua…⁶ Categ…⁷ Quali…⁸ categ…⁹
#>    <chr>         <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>  
#>  1 1             adm_un… Unique… text    integer text    NA      [INFO]… no     
#>  2 3             adm_ye… Indica… text    integer text    NA      [INFO]… no     
#>  3 4             sdc_age Partic… integer integer integer NA      [INFO]… no     
#>  4 5             sdc_ge… Gender… integer integer integer NA      [INFO]… no     
#>  5 6             phy_he… partic… decimal decimal decimal NA      [INFO]… no     
#>  6 7             phy_we… partic… decimal decimal decimal NA      [INFO]… no     
#>  7 8             phy_bmi partic… decimal decimal decimal NA      [INFO]… no     
#>  8 9             rep_pr… whethe… integer integer integer NA      [INFO]… no     
#>  9 10            rep_pr… whethe… integer integer integer NA      [INFO]… no     
#> 10 11            lsb_sm… whethe… integer integer integer NA      [INFO]… no     
#> 11 12            lsb_sm… whethe… integer integer integer NA      [INFO]… no     
#> 12 13            lsb_sm… partic… integer integer integer NA      [INFO]… no     
#> # … with 12 more variables: `Total number of observations` <dbl>,
#> #   `Nb. distinct values` <int>, `% total Valid values` <dbl>, `% NA` <dbl>,
#> #   `% Valid categorical values (if applicable)` <dbl>,
#> #   `% Missing categorical values (if applicable)` <dbl>, MIN <dbl>, Q1 <dbl>,
#> #   MEDIAN <dbl>, Q3 <dbl>, MAX <dbl>, MEAN <dbl>, and abbreviated variable
#> #   names ¹​`index in data dict.`, ²​name_var, ³​`label:en`,
#> #   ⁴​`Data Dictionary valueType`, ⁵​`Estimated dataset valueType`, …
#> 
#> 
#> $dataset_MELBOURNE_2
#> $dataset_MELBOURNE_2$`Overview of harmonization`
#> # A tibble: 13 × 2
#>    `Quality control of study specific dataset` ` `                  
#>    <chr>                                       <chr>                
#>  1 "Date"                                      "2023-03-17"         
#>  2 "Name of the dataset"                       "dataset_MELBOURNE_2"
#>  3 "    Participant identifier"                "adm_unique_id"      
#>  4 "    Variables"                             " "                  
#>  5 "        Total number of variables"         "13"                 
#>  6 "    Data type in dictionnary (valueType)"  " "                  
#>  7 "        Nb. text variables"                "1"                  
#>  8 "        Nb. date variables"                "0"                  
#>  9 "        Nb. numerical variables"           "12"                 
#> 10 "        Nb. categorical variables"         "0"                  
#> 11 "    Rows"                                  " "                  
#> 12 "        Total number of observations"      "2"                  
#> 13 "        Nb. unique participants"           "2"                  
#> 
#> $dataset_MELBOURNE_2$`Study specific Dataschema summary`
#> # A tibble: 13 × 7
#>    index name           `label:en`               value…¹ Mlstr…² Mlstr…³ Mlstr…⁴
#>    <chr> <chr>          <chr>                    <chr>   <chr>   <chr>   <chr>  
#>  1 1     adm_unique_id  Unique identification c… text    comple… "id"    id_cre…
#>  2 2     adm_study      Indicator of the survey… text    comple… "\"MEL… paste  
#>  3 3     adm_year_dce   Indicator of the survey… text    comple… "2007"  paste  
#>  4 4     sdc_age        Participant's age at ti… integer imposs… "impos… imposs…
#>  5 5     sdc_gender     Gender of the participa… integer imposs… "impos… imposs…
#>  6 6     phy_height     participant's height.    decimal imposs… "impos… imposs…
#>  7 7     phy_weight     participant's weight.    decimal imposs… "impos… imposs…
#>  8 8     phy_bmi        participant's BMI calcu… decimal imposs… "impos… imposs…
#>  9 9     rep_preg_ever  whether the participant… integer imposs… "impos… imposs…
#> 10 10    rep_preg_curr  whether the participant… integer comple… "recod… recode 
#> 11 11    lsb_smo_ever   whether the participant… integer comple… "recod… recode 
#> 12 12    lsb_smo_curr   whether the participant… integer comple… "recod… recode 
#> 13 13    lsb_smo_status participant status rega… integer comple… "recod… recode 
#> # … with abbreviated variable names ¹​valueType, ²​`Mlstr::harmo_status`,
#> #   ³​`Mlstr::algorithm`, ⁴​`Mlstr::rule_category`
#> 
#> $dataset_MELBOURNE_2$`Study specific Dataschema assessement`
#> # A tibble: 0 × 4
#> # … with 4 variables: sheet <chr>, name_col <chr>, name_var <chr>,
#> #   Quality assessment comment <chr>
#> 
#> $dataset_MELBOURNE_2$`Harmonized Dataset assessement`
#> # A tibble: 13 × 4
#>    name_var      `Quality assessment comment`                      value sugge…¹
#>    <chr>         <chr>                                             <chr> <chr>  
#>  1 adm_study     [INFO] - Unique value in the column               NA    NA     
#>  2 adm_unique_id [INFO] - refined valueType proposed               text  integer
#>  3 adm_year_dce  [INFO] - Unique value in the column               NA    NA     
#>  4 adm_year_dce  [INFO] - refined valueType proposed               text  integer
#>  5 lsb_smo_curr  [INFO] - Possible duplicated columns: lsb_smo_ev… NA    NA     
#>  6 lsb_smo_ever  [INFO] - Possible duplicated columns: lsb_smo_ev… NA    NA     
#>  7 phy_bmi       [INFO] - Empty column                             NA    NA     
#>  8 phy_height    [INFO] - Empty column                             NA    NA     
#>  9 phy_weight    [INFO] - Empty column                             NA    NA     
#> 10 rep_preg_curr [INFO] - Unique value in the column               NA    NA     
#> 11 rep_preg_ever [INFO] - Empty column                             NA    NA     
#> 12 sdc_age       [INFO] - Empty column                             NA    NA     
#> 13 sdc_gender    [INFO] - Empty column                             NA    NA     
#> # … with abbreviated variable name ¹​suggestion
#> 
#> $dataset_MELBOURNE_2$`Variables summary (all)`
#> # A tibble: 13 × 15
#>    index in da…¹ name_…² label…³ Data …⁴ Estim…⁵ Actua…⁶ Categ…⁷ Quali…⁸ categ…⁹
#>    <chr>         <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>  
#>  1 1             adm_un… Unique… text    integer text    NA      [INFO]… no     
#>  2 2             adm_st… Indica… text    text    text    NA      [INFO]… no     
#>  3 3             adm_ye… Indica… text    integer text    NA      [INFO]… no     
#>  4 4             sdc_age Partic… integer integer integer NA      [INFO]… no     
#>  5 5             sdc_ge… Gender… integer integer integer NA      [INFO]… no     
#>  6 6             phy_he… partic… decimal decimal decimal NA      [INFO]… no     
#>  7 7             phy_we… partic… decimal decimal decimal NA      [INFO]… no     
#>  8 8             phy_bmi partic… decimal decimal decimal NA      [INFO]… no     
#>  9 9             rep_pr… whethe… integer integer integer NA      [INFO]… no     
#> 10 10            rep_pr… whethe… integer integer integer NA      [INFO]… no     
#> 11 11            lsb_sm… whethe… integer integer integer NA      [INFO]… no     
#> 12 12            lsb_sm… whethe… integer integer integer NA      [INFO]… no     
#> 13 13            lsb_sm… partic… integer integer integer NA      [INFO]… no     
#> # … with 6 more variables: `Total number of observations` <dbl>,
#> #   `Nb. distinct values` <int>, `% total Valid values` <dbl>, `% NA` <dbl>,
#> #   `% Valid categorical values (if applicable)` <dbl>,
#> #   `% Missing categorical values (if applicable)` <dbl>, and abbreviated
#> #   variable names ¹​`index in data dict.`, ²​name_var, ³​`label:en`,
#> #   ⁴​`Data Dictionary valueType`, ⁵​`Estimated dataset valueType`,
#> #   ⁶​`Actual dataset valueType`, ⁷​`Categories in data dictionary`, …
#> 
#> $dataset_MELBOURNE_2$`Text variable summary`
#> # A tibble: 1 × 17
#>   index in dat…¹ name_…² label…³ Data …⁴ Estim…⁵ Actua…⁶ Categ…⁷ Quali…⁸ categ…⁹
#>   <chr>          <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>  
#> 1 2              adm_st… Indica… text    text    text    NA      [INFO]… no     
#> # … with 8 more variables: `Total number of observations` <dbl>,
#> #   `Nb. distinct values` <int>, `% total Valid values` <dbl>, `% NA` <dbl>,
#> #   `% Valid categorical values (if applicable)` <dbl>,
#> #   `% Missing categorical values (if applicable)` <dbl>,
#> #   `Most seen value(s)` <chr>, `Rarest value(s)` <chr>, and abbreviated
#> #   variable names ¹​`index in data dict.`, ²​name_var, ³​`label:en`,
#> #   ⁴​`Data Dictionary valueType`, ⁵​`Estimated dataset valueType`, …
#> 
#> $dataset_MELBOURNE_2$`Numerical variable summary`
#> # A tibble: 12 × 21
#>    index in da…¹ name_…² label…³ Data …⁴ Estim…⁵ Actua…⁶ Categ…⁷ Quali…⁸ categ…⁹
#>    <chr>         <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>  
#>  1 1             adm_un… Unique… text    integer text    NA      [INFO]… no     
#>  2 3             adm_ye… Indica… text    integer text    NA      [INFO]… no     
#>  3 4             sdc_age Partic… integer integer integer NA      [INFO]… no     
#>  4 5             sdc_ge… Gender… integer integer integer NA      [INFO]… no     
#>  5 6             phy_he… partic… decimal decimal decimal NA      [INFO]… no     
#>  6 7             phy_we… partic… decimal decimal decimal NA      [INFO]… no     
#>  7 8             phy_bmi partic… decimal decimal decimal NA      [INFO]… no     
#>  8 9             rep_pr… whethe… integer integer integer NA      [INFO]… no     
#>  9 10            rep_pr… whethe… integer integer integer NA      [INFO]… no     
#> 10 11            lsb_sm… whethe… integer integer integer NA      [INFO]… no     
#> 11 12            lsb_sm… whethe… integer integer integer NA      [INFO]… no     
#> 12 13            lsb_sm… partic… integer integer integer NA      [INFO]… no     
#> # … with 12 more variables: `Total number of observations` <dbl>,
#> #   `Nb. distinct values` <int>, `% total Valid values` <dbl>, `% NA` <dbl>,
#> #   `% Valid categorical values (if applicable)` <dbl>,
#> #   `% Missing categorical values (if applicable)` <dbl>, MIN <dbl>, Q1 <dbl>,
#> #   MEDIAN <dbl>, Q3 <dbl>, MAX <dbl>, MEAN <dbl>, and abbreviated variable
#> #   names ¹​`index in data dict.`, ²​name_var, ³​`label:en`,
#> #   ⁴​`Data Dictionary valueType`, ⁵​`Estimated dataset valueType`, …
#> 
#> 
#> $`Mlstr::pooled_harmonized_dataset`
#> $`Mlstr::pooled_harmonized_dataset`$`Overview of harmonization`
#> # A tibble: 13 × 2
#>    `Quality control of study specific dataset` ` `                              
#>    <chr>                                       <chr>                            
#>  1 "Date"                                      "2023-03-17"                     
#>  2 "Name of the dataset"                       "Mlstr::pooled_harmonized_datase…
#>  3 "    Participant identifier"                "adm_unique_id"                  
#>  4 "    Variables"                             " "                              
#>  5 "        Total number of variables"         "13"                             
#>  6 "    Data type in dictionnary (valueType)"  " "                              
#>  7 "        Nb. text variables"                "1"                              
#>  8 "        Nb. date variables"                "0"                              
#>  9 "        Nb. numerical variables"           "12"                             
#> 10 "        Nb. categorical variables"         "0"                              
#> 11 "    Rows"                                  " "                              
#> 12 "        Total number of observations"      "4"                              
#> 13 "        Nb. unique participants"           "2"                              
#> 
#> $`Mlstr::pooled_harmonized_dataset`$`Study specific Dataschema summary`
#> # A tibble: 13 × 4
#>    index name           `label:en`                                       value…¹
#>    <chr> <chr>          <chr>                                            <chr>  
#>  1 1     adm_unique_id  Unique identification code of the participant.   text   
#>  2 2     adm_study      Indicator of the survey study.                   text   
#>  3 3     adm_year_dce   Indicator of the survey data collection event.   text   
#>  4 4     sdc_age        Participant's age at time of data collection ev… integer
#>  5 5     sdc_gender     Gender of the participant.                       integer
#>  6 6     phy_height     participant's height.                            decimal
#>  7 7     phy_weight     participant's weight.                            decimal
#>  8 8     phy_bmi        participant's BMI calculation.                   decimal
#>  9 9     rep_preg_ever  whether the participant (female) has ever been … integer
#> 10 10    rep_preg_curr  whether the participant (female) is currently p… integer
#> 11 11    lsb_smo_ever   whether the participant has ever smoked          integer
#> 12 12    lsb_smo_curr   whether the participant is currently a smoker o… integer
#> 13 13    lsb_smo_status participant status regarding cigarette: never s… integer
#> # … with abbreviated variable name ¹​valueType
#> 
#> $`Mlstr::pooled_harmonized_dataset`$`Study specific Dataschema assessement`
#> # A tibble: 0 × 4
#> # … with 4 variables: sheet <chr>, name_col <chr>, name_var <chr>,
#> #   Quality assessment comment <chr>
#> 
#> $`Mlstr::pooled_harmonized_dataset`$`Harmonized Dataset assessement`
#> # A tibble: 10 × 4
#>    name_var      `Quality assessment comment`                      value sugge…¹
#>    <chr>         <chr>                                             <chr> <chr>  
#>  1 adm_study     [INFO] - Unique value in the column               NA    NA     
#>  2 adm_unique_id [INFO] - refined valueType proposed               text  integer
#>  3 adm_year_dce  [INFO] - Unique value in the column               NA    NA     
#>  4 adm_year_dce  [INFO] - refined valueType proposed               text  integer
#>  5 lsb_smo_curr  [INFO] - Possible duplicated columns: lsb_smo_ev… NA    NA     
#>  6 lsb_smo_ever  [INFO] - Possible duplicated columns: lsb_smo_ev… NA    NA     
#>  7 phy_height    [INFO] - Empty column                             NA    NA     
#>  8 phy_weight    [INFO] - Empty column                             NA    NA     
#>  9 rep_preg_curr [INFO] - Unique value in the column               NA    NA     
#> 10 rep_preg_ever [INFO] - Empty column                             NA    NA     
#> # … with abbreviated variable name ¹​suggestion
#> 
#> $`Mlstr::pooled_harmonized_dataset`$`Variables summary (all)`
#> # A tibble: 13 × 15
#>    index in da…¹ name_…² label…³ Data …⁴ Estim…⁵ Actua…⁶ Categ…⁷ Quali…⁸ categ…⁹
#>    <chr>         <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>  
#>  1 1             adm_un… Unique… text    integer text    NA      NA      no     
#>  2 2             adm_st… Indica… text    text    text    NA      [INFO]… no     
#>  3 3             adm_ye… Indica… text    integer text    NA      [INFO]… no     
#>  4 4             sdc_age Partic… integer integer integer NA      NA      no     
#>  5 5             sdc_ge… Gender… integer integer integer NA      NA      no     
#>  6 6             phy_he… partic… decimal decimal decimal NA      [INFO]… no     
#>  7 7             phy_we… partic… decimal decimal decimal NA      [INFO]… no     
#>  8 8             phy_bmi partic… decimal decimal decimal NA      NA      no     
#>  9 9             rep_pr… whethe… integer integer integer NA      [INFO]… no     
#> 10 10            rep_pr… whethe… integer integer integer NA      [INFO]… no     
#> 11 11            lsb_sm… whethe… integer integer integer NA      NA      no     
#> 12 12            lsb_sm… whethe… integer integer integer NA      NA      no     
#> 13 13            lsb_sm… partic… integer integer integer NA      NA      no     
#> # … with 6 more variables: `Total number of observations` <dbl>,
#> #   `Nb. distinct values` <int>, `% total Valid values` <dbl>, `% NA` <dbl>,
#> #   `% Valid categorical values (if applicable)` <dbl>,
#> #   `% Missing categorical values (if applicable)` <dbl>, and abbreviated
#> #   variable names ¹​`index in data dict.`, ²​name_var, ³​`label:en`,
#> #   ⁴​`Data Dictionary valueType`, ⁵​`Estimated dataset valueType`,
#> #   ⁶​`Actual dataset valueType`, ⁷​`Categories in data dictionary`, …
#> 
#> $`Mlstr::pooled_harmonized_dataset`$`Text variable summary`
#> # A tibble: 1 × 17
#>   index in dat…¹ name_…² label…³ Data …⁴ Estim…⁵ Actua…⁶ Categ…⁷ Quali…⁸ categ…⁹
#>   <chr>          <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>  
#> 1 2              adm_st… Indica… text    text    text    NA      [INFO]… no     
#> # … with 8 more variables: `Total number of observations` <dbl>,
#> #   `Nb. distinct values` <int>, `% total Valid values` <dbl>, `% NA` <dbl>,
#> #   `% Valid categorical values (if applicable)` <dbl>,
#> #   `% Missing categorical values (if applicable)` <dbl>,
#> #   `Most seen value(s)` <chr>, `Rarest value(s)` <chr>, and abbreviated
#> #   variable names ¹​`index in data dict.`, ²​name_var, ³​`label:en`,
#> #   ⁴​`Data Dictionary valueType`, ⁵​`Estimated dataset valueType`, …
#> 
#> $`Mlstr::pooled_harmonized_dataset`$`Numerical variable summary`
#> # A tibble: 12 × 21
#>    index in da…¹ name_…² label…³ Data …⁴ Estim…⁵ Actua…⁶ Categ…⁷ Quali…⁸ categ…⁹
#>    <chr>         <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>  
#>  1 1             adm_un… Unique… text    integer text    NA      NA      no     
#>  2 3             adm_ye… Indica… text    integer text    NA      [INFO]… no     
#>  3 4             sdc_age Partic… integer integer integer NA      NA      no     
#>  4 5             sdc_ge… Gender… integer integer integer NA      NA      no     
#>  5 6             phy_he… partic… decimal decimal decimal NA      [INFO]… no     
#>  6 7             phy_we… partic… decimal decimal decimal NA      [INFO]… no     
#>  7 8             phy_bmi partic… decimal decimal decimal NA      NA      no     
#>  8 9             rep_pr… whethe… integer integer integer NA      [INFO]… no     
#>  9 10            rep_pr… whethe… integer integer integer NA      [INFO]… no     
#> 10 11            lsb_sm… whethe… integer integer integer NA      NA      no     
#> 11 12            lsb_sm… whethe… integer integer integer NA      NA      no     
#> 12 13            lsb_sm… partic… integer integer integer NA      NA      no     
#> # … with 12 more variables: `Total number of observations` <dbl>,
#> #   `Nb. distinct values` <int>, `% total Valid values` <dbl>, `% NA` <dbl>,
#> #   `% Valid categorical values (if applicable)` <dbl>,
#> #   `% Missing categorical values (if applicable)` <dbl>, MIN <dbl>, Q1 <dbl>,
#> #   MEDIAN <dbl>, Q3 <dbl>, MAX <dbl>, MEAN <dbl>, and abbreviated variable
#> #   names ¹​`index in data dict.`, ²​name_var, ³​`label:en`,
#> #   ⁴​`Data Dictionary valueType`, ⁵​`Estimated dataset valueType`, …
#> 
#>