Generate R script of plots based on the 'viz_type' of the variable.

This helper function uses the visual type attributed to a variable in a data dictionary to generate R script to generate plots in a report. This function can be used to manually personalize the report parameters. The plots can use an additional variable to group each variable shown by the grouping variable.

identify_plot_type(dataset = NULL, data_dict, group_by = NULL, out = "plotly")

Arguments

dataset
A character string or tibble specifying the input dataset

data_dict
tibble of the data dictionary (automatically generated if not provided)

group_by
A character string of one column in the dataset that can be taken as a grouping column. The visual element will be grouped and displayed by this column.

out
parameter that specifies the graphical outputs expected in the report: can be either 'ggplot2' or 'plotly'. gglot2 renders static plots, plotly dynamic plots.

Value

A list of two tibbles which make up the data dictionary in Maelstrom Research format where columns plots and summary have been added to the data dictionary provided as an input.

Details

The user must run identify_viz_type first to run this function.

Examples

if (FALSE) {
# Example 1: plot R stripts for iris variables.
data_dictionary_create(iris, categories = "Species") %>%
identify_viz_type(data_dict = ., dataset = iris) %>%
  identify_plot_type(data_dict = ., dataset = iris) %>% .$Variables %>%
  select(name,viz_type, contains("plot"),contains("summary"))
}