Draw histogram plot of one (possibly grouped) column in a dataset

This function draws a histogram plot of the values of variable 'col'. Missing values can be given in input to separate non-valid from valid values, or grouped by another variable. The output can be editable (using plotly library) or static (using ggplot2 library). The R-code is also editable for refined purpose.

mlstr_plot_histogram(
  dataset = "airquality",
  col = "Ozone",
  filter = "c()",
  negate = FALSE,
  missing_values = "c()",
  out = "ggplot2",
  group_by = NULL
)

Arguments

dataset
A character string or tibble of the dataset

col
A character string of a column of interest

filter
A character string to subset the rows, applying the expressions in ... to the column values to determine which rows should be retained. It can be applied to both grouped and ungrouped data.

negate
If TRUE, return non-matching elements.

missing_values
vector listing values to exclude from valid values. Those values will not be exclud from counting, but will be separated from valid values.

out
parameter that specifies the output expected: can be either 'ggplot2', 'plotly','ggplot2-code', 'plotly-code','ggplot2-cat' or 'plotly-cat'. gglot2 renders a static plot, plotly a dynamic plot, code gives the code in a string (usable directly with eval/parse functions) and cat provides indented code in the console.

group_by
A character string of one column in the dataset that can be taken as a grouping column. the vizual element will be grouped and displayed by this column.

Value

hist plot object

Examples

if (FALSE) {
# Example 1: cat output generated as a template when no argument provided
mlstr_plot_histogram()

# Example 2: graph of Petal.Length
mlstr_plot_histogram(dataset = dataset,col = "Petal.Length",out = "ggplot2")

}