Generate a summary of the annotated data processing elements in the console

Reads annotated data processing elements to list processes, any errors, and an overview of each harmonization rule. This summary report can be used to asses harmonization and provide insight on the data structure, fields to investigate, and coherence across elements.

show_harmo_error(harmonized_dossier)

Arguments

harmonized_dossier

List of tibble(s), each of them being harmonized dataset.

Value

No object returned, the function sends messages in the console, showing errors in the harmonization process.

Details

A harmonized dossier must be a named list containing at least one data frame or data frame extension (e.g. a tibble), each of them being harmonized dataset(s). It is generally the product of applying harmonization processing to a dossier object. The name of each tibble will be use as the reference name of the dataset. A harmonized dossier has four attributes : harmonizR::class which is "harmonized_dossier" ; harmonizR::Dataschema (provided by user) ; harmonizR::data processing elements ; harmonizR::harmonized_col_id (provided by user) which refers to the column in each dataset which identifies unique combination observation/dataset. This id column name is the same across the dataset(s), the DataSchema and the data processing elements (created by using 'id_creation') and is used to initiate the process of harmonization.

Examples

{

harmonized_dossier <- DEMO_files_harmo$harmonized_dossier
show_harmo_error(harmonized_dossier)
}
#> - STATUS SUMMARY: --------------------------------------------------------
#> 
#> # A tibble: 1 × 4
#>   `Mlstr::rule_category` total_nb_errors total_nb_rules success
#>   <chr>                  <chr>           <chr>          <chr>  
#> 1 id_creation            0               1              100 %  
#>