Built-in material allowing the user to test the package with demo data

Built-in tibbles and lists allowing the user to test the package with demo material for harmonization.

DEMO_files_harmo

Format

list

A list with 15 elements used for testing the package (data frames and lists):

DEMO_data_processing_elements - final

Data processing element

DEMO_data_processing_elements - with error

Data processing element containing errors

DEMO_data_processing_elements - work in progress

Data processing element in construction

DEMO_dataschema

DataSchema used, combined with data processing elements and study-specific material

pooled_harmonized_dataset

pooled harmonized dataset for testing purpose

harmonized_dossier

harmonized dossier for testing purpose

summary_var_harmo

harmonized variables summary for testing purpose

dd_MELBOURNE_1_format_maelstrom

Data dictionary (1) of Melbourne dataset

dd_MELBOURNE_2_format_maelstrom

Data dictionary (2) of Melbourne dataset

dd_PARIS_format_maelstrom

Data dictionary of Paris dataset

dd_TOKYO_format_maelstrom

Data dictionary of Tokyo dataset

dataset_MELBOURNE_1

Dataset of Melbourne (1)

dataset_MELBOURNE_2

Dataset of Melbourne (2)

dataset_PARIS

Dataset of Paris

dataset_TOKYO

Dataset of Tokyo

...

Examples

{

 print(DEMO_files_harmo$`dataschema - final`)

}
#> $Variables
#> # A tibble: 13 × 25
#>    index name          `label:en` valueType unit  other_format individual_entity
#>    <dbl> <chr>         <chr>      <chr>     <chr> <lgl>        <lgl>            
#>  1     1 adm_unique_id Unique id… text      NA    NA           NA               
#>  2     2 adm_study     Indicator… text      NA    NA           NA               
#>  3     3 adm_year_dce  Indicator… text      NA    NA           NA               
#>  4     4 sdc_age       Participa… integer   years NA           NA               
#>  5     5 sdc_gender    Gender of… integer   NA    NA           NA               
#>  6     6 phy_height    participa… decimal   cm    NA           NA               
#>  7     7 phy_weight    participa… decimal   kg    NA           NA               
#>  8     8 phy_bmi       participa… decimal   kg/m  NA           NA               
#>  9     9 rep_preg_ever whether t… integer   NA    NA           NA               
#> 10    10 rep_preg_curr whether t… integer   NA    NA           NA               
#> 11    11 lsb_smo_ever  whether t… integer   NA    NA           NA               
#> 12    12 lsb_smo_curr  whether t… integer   NA    NA           NA               
#> 13    13 lsb_smo_stat… participa… integer   NA    NA           NA               
#> # ℹ 18 more variables: time_period <lgl>, source <lgl>, informant <lgl>,
#> #   time_collection <lgl>, wording <lgl>, measures <lgl>, procedures <lgl>,
#> #   instructions <lgl>, comments <lgl>, `Mlstr_additional::Source` <lgl>,
#> #   `Mlstr_additional::Target` <lgl>, `Mlstr_area::1` <chr>,
#> #   `Mlstr_area::1.term` <chr>, `Mlstr_area::1.scale` <chr>,
#> #   `Mlstr_area::2` <lgl>, `Mlstr_area::2.term` <lgl>, `Mlstr_area::3` <lgl>,
#> #   `Mlstr_area::3.term` <lgl>
#> 
#> $Categories
#> # A tibble: 13 × 4
#>    variable       `label:en`                            name missing
#>    <chr>          <chr>                                <dbl>   <dbl>
#>  1 sdc_gender     Male                                     1       0
#>  2 sdc_gender     Female                                   2       0
#>  3 rep_preg_ever  never pregnant                           0       0
#>  4 rep_preg_ever  pregnant once or more                    1       0
#>  5 rep_preg_curr  currently pregnant                       0       0
#>  6 rep_preg_curr  not currently pregnant                   1       0
#>  7 lsb_smo_ever   never smoked                             0       0
#>  8 lsb_smo_ever   smoked one pack of cigarette or more     1       0
#>  9 lsb_smo_curr   currently smoker                         0       0
#> 10 lsb_smo_curr   not currently smoker                     1       0
#> 11 lsb_smo_status never smoker                             0       0
#> 12 lsb_smo_status former smoker                            1       0
#> 13 lsb_smo_status current smoker                           2       0
#>