Generate a quality assessment report of the DataSchema

Assesses the content and structure of a DataSchema and reports potential issues to facilitate the assessment of input data. The report can be used to help assess data structure, presence of fields, coherence across elements, and taxonomy or data dictionary formats. This report is compatible with Excel and can be exported as an Excel spredsheet.

dataschema_evaluate(dataschema, taxonomy = NULL)

Arguments

dataschema

A list of tibble(s) representing meta data of an associated harmonized dossier.

taxonomy

A tibble identifying the scheme used for variables classification.

Value

A list of tibbles of report for the DataSchema.

Details

A DataSchema defines the harmonized variables to be generated, representing meta data of an associated harmonized dossier. It must be a list of data frame like objects with elements named 'Variables' (required) and 'Categories' (if any). The 'Variables' element must contain at least the 'name' column, and the 'Categories' element must contain at least the 'variable' and 'name' columns to be usable in any function. To be considered as a minimum workable DataSchema, in 'Variables' the 'name' column must also have unique and non-null entries, and in 'Categories' the combination of 'variable' and 'name' columns must also be unique'.

A taxonomy is classification scheme that can be defined for variable attributes. If defined, a taxonomy must be a data frame like object. It must be compatible with (and is generally extracted from) an Opal environment. To work with certain functions, a valid taxonomy must contain at least the columns 'taxonomy', 'vocabulary', and 'terms'. In addition, the taxonomy may follow Maelstrom research taxonomy, and its content can be evaluated accordingly, such as naming convention restriction, tagging elements, or scales, which are specific to Maelstrom Research. In this particular case, the tibble must also contain 'vocabulary_short', 'taxonomy_scale', 'vocabulary_scale' and 'term_scale' to work with some specific functions.

Examples

{

# You can use our demonstration files to run examples
library(madshapR) # import DEMO_files

dataschema <- DEMO_files$`dataschema - final`
dataschema_evaluate(dataschema)

}
#> - DATA DICTIONARY ASSESSMENT: dataschema --------------
#>     Assess the standard adequacy of naming
#>     Assess the uniqueness of variable names
#>     Assess the presence of possible duplicated columns
#>     Assess the presence of empty rows in the data dictionary
#>     Assess the presence of empty columns in the data dictionary
#>     Assess the presence of categories not in the data dictionary
#>     Assess the completion of `label(:xx)` column in 'Variables'
#>     Assess the `valueType` column in 'Variables'
#>     Assess presence and completion of `label(:xx)` column in 'Categories'
#>     Assess the logical values of missing column in Categories
#>     Generate report
#> 
#>   - WARNING MESSAGES (if any): --------------------------------------------
#> 
#> $`Study-specific Dataschema summary`
#> # A tibble: 13 × 31
#>    index name     `label:en` valueType `Categories::project` `Categories::table`
#>    <int> <chr>    <chr>      <chr>     <chr>                 <chr>              
#>  1     1 adm_uni… Unique id… text       NA                    NA                
#>  2     2 adm_stu… Indicator… text       NA                    NA                
#>  3     3 adm_yea… Indicator… text       NA                    NA                
#>  4     4 sdc_age  Participa… integer    NA                    NA                
#>  5     5 sdc_gen… Gender of… integer   "1 = DEMO ; \n2 = DE… "1 = harmonized_st…
#>  6     6 phy_hei… participa… decimal    NA                    NA                
#>  7     7 phy_wei… participa… decimal    NA                    NA                
#>  8     8 phy_bmi  participa… decimal    NA                    NA                
#>  9     9 rep_pre… whether t… integer   "0 = DEMO ; \n1 = DE… "0 = harmonized_st…
#> 10    10 rep_pre… whether t… integer   "0 = DEMO ; \n1 = DE… "0 = harmonized_st…
#> 11    11 lsb_smo… whether t… integer   "0 = DEMO ; \n1 = DE… "0 = harmonized_st…
#> 12    12 lsb_smo… whether t… integer   "0 = DEMO ; \n1 = DE… "0 = harmonized_st…
#> 13    13 lsb_smo… participa… integer   "0 = DEMO ; \n1 = DE… "0 = harmonized_st…
#> # ℹ 25 more variables: `Categories::label:en` <chr>,
#> #   `Categories::missing` <chr>, project <chr>, table <chr>, unit <chr>,
#> #   other_format <chr>, individual_entity <chr>, time_period <chr>,
#> #   source <chr>, informant <chr>, time_collection <chr>, wording <chr>,
#> #   measures <chr>, procedures <chr>, instructions <chr>, comments <chr>,
#> #   `Mlstr_additional::Source` <chr>, `Mlstr_additional::Target` <chr>,
#> #   `Mlstr_area::1` <chr>, `Mlstr_area::1.term` <chr>, …
#> 
#> $`Study-specific Dataschema assessement`
#> # A tibble: 17 × 4
#>    sheet     name_col                 name_var `Quality assessment comment`
#>    <chr>     <chr>                    <chr>    <chr>                       
#>  1 Variables Mlstr_additional::Source NA       [INFO] - Empty column       
#>  2 Variables Mlstr_additional::Target NA       [INFO] - Empty column       
#>  3 Variables Mlstr_area::2            NA       [INFO] - Empty column       
#>  4 Variables Mlstr_area::2.term       NA       [INFO] - Empty column       
#>  5 Variables Mlstr_area::3            NA       [INFO] - Empty column       
#>  6 Variables Mlstr_area::3.term       NA       [INFO] - Empty column       
#>  7 Variables comments                 NA       [INFO] - Empty column       
#>  8 Variables individual_entity        NA       [INFO] - Empty column       
#>  9 Variables informant                NA       [INFO] - Empty column       
#> 10 Variables instructions             NA       [INFO] - Empty column       
#> 11 Variables measures                 NA       [INFO] - Empty column       
#> 12 Variables other_format             NA       [INFO] - Empty column       
#> 13 Variables procedures               NA       [INFO] - Empty column       
#> 14 Variables source                   NA       [INFO] - Empty column       
#> 15 Variables time_collection          NA       [INFO] - Empty column       
#> 16 Variables time_period              NA       [INFO] - Empty column       
#> 17 Variables wording                  NA       [INFO] - Empty column       
#>