Generate a report of a harmonized dossier (list of harmonized datasets)

Assesses the content and structure of a harmonized dossier object (list of harmonized datasets) and reports possible issues in the datasets and data dictionaries to facilitate assessment of input data. The report can be used to help assess data structure, presence of fields, coherence across elements, and taxonomy or data dictionary formats. This report is compatible with Excel and can be exported as an Excel spredsheet.

harmonized_dossier_evaluate(harmonized_dossier, taxonomy = NULL)

Arguments

harmonized_dossier

List of tibble(s), each of them being harmonized dataset.

taxonomy

A tibble identifying the scheme used for variables classification.

Value

A list of report(s), each of them being tibble(s) ('Overview and summary) for each harmonized dataset.

Details

A harmonized dossier must be a named list containing at least one data frame or data frame extension (e.g. a tibble), each of them being harmonized dataset(s). It is generally the product of applying harmonization processing to a dossier object. The name of each tibble will be use as the reference name of the dataset. A harmonized dossier has four attributes : harmonizR::class which is ""harmonized_dossier"" ; harmonizR::Dataschema (provided by user) ; harmonizR::data processing elements ; harmonizR::harmonized_col_id (provided by user) which refers to the column in each dataset which identifies unique combination observation/dataset. This id column name is the same across the dataset(s), the DataSchema and the data processing elements (created by using 'id_creation') and is used to initiate the process of harmonization.

A taxonomy is classification scheme that can be defined for variable attributes. If defined, a taxonomy must be a data frame like object. It must be compatible with (and is generally extracted from) an Opal environment. To work with certain functions, a valid taxonomy must contain at least the columns 'taxonomy', 'vocabulary', and 'terms'. In addition, the taxonomy may follow Maelstrom research taxonomy, and its content can be evaluated accordingly, such as naming convention restriction, tagging elements, or scales, which are specific to Maelstrom Research. In this particular case, the tibble must also contain 'vocabulary_short', 'taxonomy_scale', 'vocabulary_scale' and 'term_scale' to work with some specific functions.

Examples

{

# You can use our demonstration files to run examples
library(madshapR) # import DEMO_files
library(fabR)     # silently_run
library(lubridate)

dataset_MELBOURNE_1 <- DEMO_files$dataset_MELBOURNE_1
dataset_MELBOURNE_2 <- DEMO_files$dataset_MELBOURNE_2
dossier <- dossier_create(list(dataset_MELBOURNE_1, dataset_MELBOURNE_2))

dataschema <- DEMO_files$`dataschema - final`

data_proc_elem <- 
  DEMO_files$`data_processing_elements - with error`$`dataProcessing`
  
# perform harmonization 
harmonized_dossier <- 
  silently_run(harmo_process(dossier,dataschema,data_proc_elem))

# evaluate harmonization
harmonized_dossier_evaluate(harmonized_dossier)

}
#> - DOSSIER ASSESSMENT: ----------------------------------------------------
#> - DATA DICTIONARY ASSESSMENT: data_dict --------------
#>     Assess the standard adequacy of naming
#>     Assess the uniqueness of variable names
#>     Assess the presence of possible duplicated columns
#>     Assess the presence of empty rows in the data dictionary
#>     Assess the presence of empty columns in the data dictionary
#>     Assess the presence of categories not in the data dictionary
#>     Assess the completion of `label(:xx)` column in 'Variables'
#>     Assess the `valueType` column in 'Variables'
#>     Assess presence and completion of `label(:xx)` column in 'Categories'
#>     Assess the logical values of missing column in Categories
#>     Generate report
#> 
#>     The data dictionary contains no error/warning.
#> 
#>   - WARNING MESSAGES (if any): --------------------------------------------
#> 
#> - DATASET ASSESSMENT: dataset_MELBOURNE_1 (empty dataset) --------------------------
#>     Assess the standard adequacy of naming
#>     Assess the presence of variable names both in dataset and data dictionary
#>     Assess the presence of possible duplicated variable in the dataset
#>     Assess the presence of duplicated participants in the dataset
#>     Assess the presence of empty rows in the data dictionary
#>     Assess the presence all NA(s) of columns in the data dictionary
#>     Assess the presence of categories not in the data dictionary
#>     Assess the `valueType` comparison in dataset and data dictionary
#>     Generate report
#> 
#>   - WARNING MESSAGES (if any): -------------------------------------------------
#>     
#> - DATA DICTIONARY ASSESSMENT: data_dict --------------
#>     Assess the standard adequacy of naming
#>     Assess the uniqueness of variable names
#>     Assess the presence of possible duplicated columns
#>     Assess the presence of empty rows in the data dictionary
#>     Assess the presence of empty columns in the data dictionary
#>     Assess the presence of categories not in the data dictionary
#>     Assess the completion of `label(:xx)` column in 'Variables'
#>     Assess the `valueType` column in 'Variables'
#>     Assess presence and completion of `label(:xx)` column in 'Categories'
#>     Assess the logical values of missing column in Categories
#>     Generate report
#> 
#>     The data dictionary contains no error/warning.
#> 
#>   - WARNING MESSAGES (if any): --------------------------------------------
#> 
#> - DATASET ASSESSMENT: dataset_MELBOURNE_2 (empty dataset) --------------------------
#>     Assess the standard adequacy of naming
#>     Assess the presence of variable names both in dataset and data dictionary
#>     Assess the presence of possible duplicated variable in the dataset
#>     Assess the presence of duplicated participants in the dataset
#>     Assess the presence of empty rows in the data dictionary
#>     Assess the presence all NA(s) of columns in the data dictionary
#>     Assess the presence of categories not in the data dictionary
#>     Assess the `valueType` comparison in dataset and data dictionary
#>     Generate report
#> 
#>   - WARNING MESSAGES (if any): -------------------------------------------------
#>     
#> $dataset_MELBOURNE_1
#> $dataset_MELBOURNE_1$`Study-specific Dataschema summary`
#> # A tibble: 13 × 18
#>    index name  `label:en` valueType `Categories::label:en` `Categories::missing`
#>    <int> <chr> <chr>      <chr>     <chr>                  <chr>                
#>  1     1 adm_… Unique id… text       NA                     NA                  
#>  2     2 adm_… Indicator… text       NA                     NA                  
#>  3     3 adm_… Indicator… text       NA                     NA                  
#>  4     4 sdc_… Participa… integer    NA                     NA                  
#>  5     5 sdc_… Gender of… integer   "1 = Male ; \n2 = Fem… "1 = FALSE ; \n2 = F…
#>  6     6 phy_… participa… decimal    NA                     NA                  
#>  7     7 phy_… participa… decimal    NA                     NA                  
#>  8     8 phy_… participa… decimal    NA                     NA                  
#>  9     9 rep_… whether t… integer   "0 = never pregnant ;… "0 = FALSE ; \n1 = F…
#> 10    10 rep_… whether t… integer   "0 = currently pregna… "0 = FALSE ; \n1 = F…
#> 11    11 lsb_… whether t… integer   "0 = never smoked ; \… "0 = FALSE ; \n1 = F…
#> 12    12 lsb_… whether t… integer   "0 = currently smoker… "0 = FALSE ; \n1 = F…
#> 13    13 lsb_… participa… integer   "0 = never smoker ; \… "0 = FALSE ; \n1 = F…
#> # ℹ 12 more variables: `Categories::project` <chr>, `Categories::table` <chr>,
#> #   project <chr>, table <chr>, `Mlstr_area::1` <chr>,
#> #   `Mlstr_area::1.term` <chr>, `Mlstr::rule_category` <chr>,
#> #   `Mlstr::algorithm` <chr>, `harmonizR::r_script` <chr>,
#> #   `Mlstr::status` <chr>, unit <chr>, `Mlstr_area::1.scale` <chr>
#> 
#> $dataset_MELBOURNE_1$`Harmonized Dataset assessment`
#> # A tibble: 13 × 5
#>    `index in data dict.` name           Quality assessment co…¹ value suggestion
#>                    <int> <chr>          <chr>                   <chr> <chr>     
#>  1                     1 adm_unique_id  [INFO] - Empty column   NA    NA        
#>  2                     2 adm_study      [INFO] - Empty column   NA    NA        
#>  3                     3 adm_year_dce   [INFO] - Empty column   NA    NA        
#>  4                     4 sdc_age        [INFO] - Empty column   NA    NA        
#>  5                     5 sdc_gender     [INFO] - Empty column   NA    NA        
#>  6                     6 phy_height     [INFO] - Empty column   NA    NA        
#>  7                     7 phy_weight     [INFO] - Empty column   NA    NA        
#>  8                     8 phy_bmi        [INFO] - Empty column   NA    NA        
#>  9                     9 rep_preg_ever  [INFO] - Empty column   NA    NA        
#> 10                    10 rep_preg_curr  [INFO] - Empty column   NA    NA        
#> 11                    11 lsb_smo_ever   [INFO] - Empty column   NA    NA        
#> 12                    12 lsb_smo_curr   [INFO] - Empty column   NA    NA        
#> 13                    13 lsb_smo_status [INFO] - Empty column   NA    NA        
#> # ℹ abbreviated name: ¹​`Quality assessment comment`
#> 
#> 
#> $dataset_MELBOURNE_2
#> $dataset_MELBOURNE_2$`Study-specific Dataschema summary`
#> # A tibble: 13 × 18
#>    index name  `label:en` valueType `Categories::label:en` `Categories::missing`
#>    <int> <chr> <chr>      <chr>     <chr>                  <chr>                
#>  1     1 adm_… Unique id… text       NA                     NA                  
#>  2     2 adm_… Indicator… text       NA                     NA                  
#>  3     3 adm_… Indicator… text       NA                     NA                  
#>  4     4 sdc_… Participa… integer    NA                     NA                  
#>  5     5 sdc_… Gender of… integer   "1 = Male ; \n2 = Fem… "1 = FALSE ; \n2 = F…
#>  6     6 phy_… participa… decimal    NA                     NA                  
#>  7     7 phy_… participa… decimal    NA                     NA                  
#>  8     8 phy_… participa… decimal    NA                     NA                  
#>  9     9 rep_… whether t… integer   "0 = never pregnant ;… "0 = FALSE ; \n1 = F…
#> 10    10 rep_… whether t… integer   "0 = currently pregna… "0 = FALSE ; \n1 = F…
#> 11    11 lsb_… whether t… integer   "0 = never smoked ; \… "0 = FALSE ; \n1 = F…
#> 12    12 lsb_… whether t… integer   "0 = currently smoker… "0 = FALSE ; \n1 = F…
#> 13    13 lsb_… participa… integer   "0 = never smoker ; \… "0 = FALSE ; \n1 = F…
#> # ℹ 12 more variables: `Categories::project` <chr>, `Categories::table` <chr>,
#> #   project <chr>, table <chr>, `Mlstr_area::1` <chr>,
#> #   `Mlstr_area::1.term` <chr>, `Mlstr::rule_category` <chr>,
#> #   `Mlstr::algorithm` <chr>, `harmonizR::r_script` <chr>,
#> #   `Mlstr::status` <chr>, unit <chr>, `Mlstr_area::1.scale` <chr>
#> 
#> $dataset_MELBOURNE_2$`Harmonized Dataset assessment`
#> # A tibble: 13 × 5
#>    `index in data dict.` name           Quality assessment co…¹ value suggestion
#>                    <int> <chr>          <chr>                   <chr> <chr>     
#>  1                     1 adm_unique_id  [INFO] - Empty column   NA    NA        
#>  2                     2 adm_study      [INFO] - Empty column   NA    NA        
#>  3                     3 adm_year_dce   [INFO] - Empty column   NA    NA        
#>  4                     4 sdc_age        [INFO] - Empty column   NA    NA        
#>  5                     5 sdc_gender     [INFO] - Empty column   NA    NA        
#>  6                     6 phy_height     [INFO] - Empty column   NA    NA        
#>  7                     7 phy_weight     [INFO] - Empty column   NA    NA        
#>  8                     8 phy_bmi        [INFO] - Empty column   NA    NA        
#>  9                     9 rep_preg_ever  [INFO] - Empty column   NA    NA        
#> 10                    10 rep_preg_curr  [INFO] - Empty column   NA    NA        
#> 11                    11 lsb_smo_ever   [INFO] - Empty column   NA    NA        
#> 12                    12 lsb_smo_curr   [INFO] - Empty column   NA    NA        
#> 13                    13 lsb_smo_status [INFO] - Empty column   NA    NA        
#> # ℹ abbreviated name: ¹​`Quality assessment comment`
#> 
#>