Generate a web base application visual report of a harmonization project

Generates a visual report for a harmonized dossier in an HTML bookdown document, showing descriptive statistics for each harmonized variable to facilitate the assessment of input data. The types of statistics and figures are generated according to valueType. This report can be used to assess the data structure, fields to investigate , or coherence across elements and taxonomy (if provided). The summaries and figures associated provide an overview of dataset composition, with variable distributions and descriptive statistics.

harmonized_dossier_visualize(
  harmonized_dossier,
  to,
  taxonomy = NULL,
  .keep_files = TRUE,
  .summary_var = NULL
)

Arguments

harmonized_dossier

List of tibble(s), each of them being harmonized dataset.

to

**

taxonomy

A data-frame or data-frame extension (e.g. a tibble), identifying the scheme used for variables classification as a tibble.

.keep_files

whether to keep the R-markdown files. TRUE by default.

.summary_var

A summary list which provides summary of the variables ( If provided, the function will not do this action internally, enhancing the speed of completion)

Value

A bookdown folder containing files in the specified output folder. To open the file in browser, open 'docs/index.html'. Or use open_visual_report()

Details

A harmonized dossier must be a named list containing at least one data-frame or data-frame extension (e.g. a tibble), each of them being harmonized dataset(s). The name of each tibble will be use as the reference name of the dataset. A harmonized dossier has four attributes : harmonizR::class which is "harmonized_dossier" ; harmonizR::Dataschema (provided by user) ; harmonizR::data processing elements ; harmonizR::harmonized_col_id (provided by user) which refers to the column in each dataset which identifies unique combination observation/dataset. This id column name is the same across the dataset(s), the DataSchema and the data processing elements (created by using 'id_creation') and is used to initiate the process of harmonization.

A taxonomy must be a data-frame or data-frame extension (e.g. a tibble). The taxonomy must be compatible with (and generally extracted from) an Opal environment, and must contain at least 'taxonomy', 'vocabulary' and 'terms' to work with some specific functions. In addition, the taxonomy may follow Maelstrom research standards, and its content can be evaluated accordingly, such as naming convention restriction, tagging elements, or scales, which are specific to Maelstrom Research. In this particular case, the tibble must also contain 'vocabulary_short', 'taxonomy_scale', 'vocabulary_scale' and 'term_scale' to work with some specific functions.

See also

Examples

if (FALSE) {

# You can use our demonstration files to run examples
library(madshapR) # import DEMO_files
library(fabR) # silently_run

dataset_MELBOURNE_1 <- DEMO_files$dataset_MELBOURNE_1
dataset_MELBOURNE_2 <- DEMO_files$dataset_MELBOURNE_2
dossier <- dossier_create(list(dataset_MELBOURNE_1, dataset_MELBOURNE_2))

dataschema <- DEMO_files$`dataschema - final`

data_proc_elem <- 
  DEMO_files$`data_processing_elements - final`$`dataProcessing`
  
# perform harmonization 
harmonized_dossier <- 
  silently_run(harmo_process(dossier,dataschema,data_proc_elem))

# generate viz report
to = tempdir()

# Silencing the process, remove silently_run() to see the full process in 
# the console. 
silently_run(harmonized_dossier_visualize(harmonized_dossier, to=tempdir()))
 
open_visual_report(to)

}