Generate summary of data processing elements to display in the console

Reads annotated data processing elements to list processes, errors, and an overview of each rule of harmonization. This summary report can be used to assist the user in the assessment of harmonization, providing insights on the data structure, fields to investigate, coherence across elements or standard evaluation.

show_harmo_error(harmonization_output)

Arguments

harmonization_output

A tibble, identifying of the annotated data processing elements. These elements are the centralised decision making to perform harmonization of each study-specific dataset.

Value

No object returned, the function sends messages in the console, showing errors in the harmonization process

Details

A data processing element must be a list of at least one data frame or data frame extension (e.g. a tibble) named 'dataProcess'. These elements are the centralised decision making to perform harmonization of each study-specific dataset. The tibble must contain certain columns which participate to the process, including the 'rule_category' and 'harmo_rule'. In addition, the data processing element may follow Maelstrom research standards, and its content can be evaluated accordingly, such as naming convention restriction, columns like 'valueType', and 'harmo status details', and/or any harmonization taxonomy provided.

Examples

{

library(madshapR)
library(lubridate)
library(dplyr)

dataset_MELBOURNE_1 <- DEMO_files$dataset_MELBOURNE_1
dataset_MELBOURNE_2 <- DEMO_files$dataset_MELBOURNE_2
study <-
  study_create(list(dataset_MELBOURNE_1, dataset_MELBOURNE_2)) %>%
  lapply(. %>% slice(1:2))

dataschema <-
  list(Variables =
        DEMO_files$`dataschema - final`$Variables[
           c('name','label:en','valueType')])

data_procelem <- DEMO_files$`data_processing_elements - with error`
harmonization_output <- harmo_process(study,dataschema, data_procelem)

show_harmo_error(harmonization_output)
}
#> - DATA PROCESSING ELEMENTS: ------------------------------------------------
#> 
#> --harmonization of : dataset_MELBOURNE_1 ------------------------------
#>     processing 1/13 : adm_unique_id     id created
#>     processing 2/13 : adm_study         complete
#>     processing 3/13 : adm_year_dce      complete
#>     processing 4/13 : sdc_age           complete
#>     processing 5/13 : sdc_gender        **ERROR**
#>     processing 6/13 : phy_height        impossible
#>     processing 7/13 : phy_weight        impossible
#>     processing 8/13 : phy_bmi           complete
#>     processing 9/13 : rep_preg_ever     impossible
#>     processing 10/13 : rep_preg_curr    impossible
#>     processing 11/13 : lsb_smo_ever     impossible
#>     processing 12/13 : lsb_smo_curr     impossible
#>     processing 13/13 : lsb_smo_status   impossible
#> 
#> --harmonization of : dataset_MELBOURNE_2 ------------------------------
#>     processing 1/13 : adm_unique_id     id created
#>     processing 2/13 : adm_study         complete
#>     processing 3/13 : adm_year_dce      complete
#>     processing 4/13 : sdc_age           impossible
#>     processing 5/13 : sdc_gender        impossible
#>     processing 6/13 : phy_height        impossible
#>     processing 7/13 : phy_weight        impossible
#>     processing 8/13 : phy_bmi           impossible
#>     processing 9/13 : rep_preg_ever     impossible
#>     processing 10/13 : rep_preg_curr    complete
#>     processing 11/13 : lsb_smo_ever     complete
#>     processing 12/13 : lsb_smo_curr     **ERROR**
#>     processing 13/13 : lsb_smo_status   complete
#> 
#> 
#> ------------------------------------------------------------------------------
#> 
#> Your harmonization process contains errors. The tables affected have not been
#> generated to avoid any version confusion. Correct your data processing element
#> file and re-run the process.
#> 
#> 
#> Useful tip: If you identified errors and want to correct them later, you can specify
#> 'undetermined' in the column 'harmo_rule' of your problematic data processing
#> element(s). Such data processing element will not be processed.
#> 
#> 
#> - ERROR STATUS DETAILS: ---------------------------------------------------
#> Here is the list of the errors encountered in the process of harmonization:
#> -----------------------------------------------------------------------------
#> sdc_gender in dataset_MELBOURNE_1 
#> 
#> 
#> Pseudo Code:
#> direct_mapping
#> 
#> R Script:
#> input_table %>% 
#>   mutate(
#>   sdc_gender = gendr) %>%
#>   select(1,sdc_gender)
#> 
#> R Error:
#> Error in mutate(., sdc_gender = gendr) : 
#>   ℹ In argument: `sdc_gender = gendr`.
#> Caused by error:
#> ! object 'gendr' not found
#> 
#> -----------------------------------------------------------------------------
#> lsb_smo_curr in dataset_MELBOURNE_2 
#> 
#> 
#> Pseudo Code:
#> recode(1 = 0 ; 3 = 0 ; 2 = 1  ELSE = NA)
#> 
#> R Script:
#> input_table %>% 
#>   select(1, smo_status) %>% 
#>   mutate(
#>   lsb_smo_curr = car::recode(
#>       var = .$smo_status ,
#>       recodes =  '1 = 0 ; 3 = 0 ; 2 = 1  else  = NA')) %>%
#>   select(1,lsb_smo_curr)
#> 
#> R Error:
#> Error in mutate(., lsb_smo_curr = car::recode(var = .$smo_status, recodes = "1 = 0 ; 3 = 0 ; 2 = 1  else  = NA")) : 
#>   ℹ In argument: `lsb_smo_curr = car::recode(var = .$smo_status, recodes =
#>   "1 = 0 ; 3 = 0 ; 2 = 1 else = NA")`.
#> Caused by error in `car::recode()`:
#> ! 
#>   in recode term:  2 = 1  else  = NA
#>   message: Error in parse(text = strsplit(term, to.value)[[1]][2]) : 
#>   <text>:1:5: unexpected 'else'
#> 1:  1  else
#>         ^
#> 
#> - STATUS SUMMARY: --------------------------------------------------------
#> 
#> # A tibble: 5 × 4
#>   rule_category  total_nb_errors total_nb_rules success
#>   <chr>          <chr>           <chr>          <chr>  
#> 1 direct_mapping 1               2              50 %   
#> 2 recode         1               5              80 %   
#> 3 id_creation    0               2              100 %  
#> 4 impossible     0               13             100 %  
#> 5 paste          0               4              100 %  
#>