Generate a summary of the annotated data processing elements in the console

Reads annotated data processing elements to list processes, any errors, and an overview of each harmonization rule. This summary report can be used to asses harmonization and provide insight on the data structure, fields to investigate, and coherence across elements.

show_harmo_error(harmonized_dossier)

Arguments

harmonized_dossier

List of tibble(s), each of them being harmonized dataset.

Value

No object returned, the function sends messages in the console, showing errors in the harmonization process

Details

A harmonized dossier must be a named list containing at least one data-frame or data-frame extension (e.g. a tibble), each of them being harmonized dataset(s). The name of each tibble will be use as the reference name of the dataset. A harmonized dossier has four attributes : harmonizR::class which is "harmonized_dossier" ; harmonizR::Dataschema (provided by user) ; harmonizR::data processing elements ; harmonizR::harmonized_col_id (provided by user) which refers to the column in each dataset which identifies unique combination observation/dataset. This id column name is the same across the dataset(s), the DataSchema and the data processing elements (created by using 'id_creation') and is used to initiate the process of harmonization.

Examples

{

# You can use our demonstration files to run examples
library(madshapR) # import DEMO_files
library(lubridate)

dataset_MELBOURNE_1 <- DEMO_files$dataset_MELBOURNE_1
dataset_MELBOURNE_2 <- DEMO_files$dataset_MELBOURNE_2
dossier <- dossier_create(list(dataset_MELBOURNE_1, dataset_MELBOURNE_2))

dataschema <- DEMO_files$`dataschema - final`

data_proc_elem <- 
  DEMO_files$`data_processing_elements - with error`$`dataProcessing`
  
# perform harmonization 
harmonized_dossier <- harmo_process(dossier,dataschema,data_proc_elem)

show_harmo_error(harmonized_dossier)
}
#> - DATA PROCESSING ELEMENTS: ------------------------------------------------
#> 
#> --harmonization of : dataset_MELBOURNE_1 ------------------------------
#>     processing 1/13 : adm_unique_id     id created
#>     processing 2/14 : adm_study         complete
#>     processing 3/14 : adm_year_dce      complete
#>     processing 4/14 : sdc_age           complete
#>     processing 5/14 : sdc_gender        **ERROR**
#>     processing 6/14 : phy_height        impossible
#>     processing 7/14 : phy_weight        impossible
#>     processing 8/14 : phy_bmi           complete
#>     processing 9/14 : rep_preg_ever     impossible
#>     processing 10/14 : rep_preg_curr    impossible
#>     processing 11/14 : lsb_smo_ever     impossible
#>     processing 12/14 : lsb_smo_curr     impossible
#>     processing 13/14 : lsb_smo_status   impossible
#> 
#> --harmonization of : dataset_MELBOURNE_2 ------------------------------
#>     processing 1/13 : adm_unique_id     id created
#>     processing 2/14 : adm_study         complete
#>     processing 3/14 : adm_year_dce      complete
#>     processing 4/14 : sdc_age           impossible
#>     processing 5/14 : sdc_gender        impossible
#>     processing 6/14 : phy_height        impossible
#>     processing 7/14 : phy_weight        impossible
#>     processing 8/14 : phy_bmi           impossible
#>     processing 9/14 : rep_preg_ever     impossible
#>     processing 10/14 : rep_preg_curr    complete
#>     processing 11/14 : lsb_smo_ever     complete
#>     processing 12/14 : lsb_smo_curr     **ERROR**
#>     processing 13/14 : lsb_smo_status   complete
#> 
#> 
#> - CREATION OF STUDY-SPECIFIC DATASCHEMA: ----------------------------------
#> 
#> dataset_MELBOURNE_1 : done
#> dataset_MELBOURNE_2 : done
#> 
#> 
#> ------------------------------------------------------------------------------
#> 
#> Your harmonization process contains errors. The tables with error(s) have not 
#> been generated to avoid any version confusion. Correct your data processing 
#> elements file and re-run the process.
#> 
#> 
#> Useful tip: If you identified errors and want to correct them later, you can specify
#> 'undetermined' in the column 'Mlstr::algorithm' of your problematic 
#> data processing elements(s). Such data processing elements will not be 
#> processed.
#> 
#> - WARNING MESSAGES (if any): ----------------------------------------------
#> 
#> 
#> 
#> - ERROR STATUS DETAILS: ---------------------------------------------------
#> Here is the list of the errors encountered in the process of harmonization:
#> -----------------------------------------------------------------------------
#> sdc_gender in dataset_MELBOURNE_1 
#> 
#> 
#> Pseudo Code:
#> direct_mapping
#> 
#> R Script:
#> dataset_MELBOURNE_1 %>% 
#>   mutate(
#>   sdc_gender = gendr) %>%
#>   select(1,sdc_gender)
#> 
#> R Error:
#> Error in mutate(., sdc_gender = gendr) : 
#>   ℹ In argument: `sdc_gender = gendr`.
#> Caused by error:
#> ! object 'gendr' not found
#> 
#> -----------------------------------------------------------------------------
#> lsb_smo_curr in dataset_MELBOURNE_2 
#> 
#> 
#> Pseudo Code:
#> recode(1 = 0 ; 3 = 0 ; 2 = 1  ELSE = NA)
#> 
#> R Script:
#> dataset_MELBOURNE_2 %>% 
#>   select(1, smo_status) %>% 
#>   mutate(
#>   lsb_smo_curr = car::recode(
#>       var = .$smo_status ,
#>       recodes =  '1 = 0 ; 3 = 0 ; 2 = 1  else  = NA')) %>%
#>   select(1,lsb_smo_curr)
#> 
#> R Error:
#> Error in mutate(., lsb_smo_curr = car::recode(var = .$smo_status, recodes = "1 = 0 ; 3 = 0 ; 2 = 1  else  = NA")) : 
#>   ℹ In argument: `lsb_smo_curr = car::recode(var = .$smo_status, recodes =
#>   "1 = 0 ; 3 = 0 ; 2 = 1 else = NA")`.
#> Caused by error in `car::recode()`:
#> ! 
#>   in recode term:  2 = 1  else  = NA
#>   message: Error in parse(text = strsplit(term, to.value)[[1]][2]) : 
#>   <text>:1:5: unexpected 'else'
#> 1:  1  else
#>         ^
#> 
#> - STATUS SUMMARY: --------------------------------------------------------
#> 
#> # A tibble: 5 × 4
#>   `Mlstr::rule_category` total_nb_errors total_nb_rules success
#>   <chr>                  <chr>           <chr>          <chr>  
#> 1 direct_mapping         1               2              50 %   
#> 2 recode                 1               5              80 %   
#> 3 id_creation            0               2              100 %  
#> 4 impossible             0               13             100 %  
#> 5 paste                  0               4              100 %  
#>